builtin/mv: fix out of bounds write
When commit a88c915 (mv: move submodules using a gitfile, 2013-07-30)
added the submodule_gitfile array, it was not added to the block that
enlarges the arrays when we are moving a directory so that we do not
have to worry about it being a directory when we perform the actual
move.  After this, the loop continues over the enlarged set of sources.

Since we assume that submodule_gitfile has size argc, if any of the
items in the source directory are submodules we are guaranteed to write
beyond the end of submodule_gitfile.

Fix this by realloc'ing submodule_gitfile at the same time as the other
arrays.

Reported-by: Guillaume Gelin <contact@ramnes.eu>
Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7308e2b10f77deca03aff4765985b528713d2eda [formerly 7308e2b10f77deca03aff4765985b528713d2eda [formerly 89ccc1b09cf4004e6129c66def42b47206ed6b5f]]
Former-commit-id: 61ccfe199dfb39bc05498639c7aaa86c72bdd895
Former-commit-id: 568f62b75bea1d056ac7601f2e9c6491c96f4a48