clean: simplify dir/not-dir logic
When we get a list of paths from read_directory, we further
prune it to create the final list of items to remove. The
code paths for directories and non-directories repeat the
same "add to list" code.

This patch restructures the code so that we don't repeat
ourselves. Also, by following a "if (condition) continue"
pattern like the pathspec check above, it makes it more
obvious that the conditional is about excluding directories
under certain circumstances.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 28741a0bcb5306edef8ae78cf5572242b365092b [formerly 28741a0bcb5306edef8ae78cf5572242b365092b [formerly 1f2e1088871e777355025a84f80a2d8b5cb04c06]]
Former-commit-id: ed115b2357163efc58608a05829f453d3578b082
Former-commit-id: ada31dbaa6dfe40e39510def84db4cec4375b522