clean: respect pathspecs with "-d"
git-clean uses read_directory to fill in a `struct dir` with
potential hits. However, read_directory does not actually
check against our pathspec. It uses a simplified version
that may turn up false positives. As a result, we need to
check that any hits match our pathspec. We do so reliably
for non-directories. For directories, if "-d" is not given
we check that the pathspec matched exactly (i.e., we are
even stricter, and require an explicit "git clean foo" to
clean "foo/"). But if "-d" is given, rather than relaxing
the exact match to allow a recursive match, we do not check
the pathspec at all.

This regression was introduced in 113f10f (Make git-clean a
builtin, 2007-11-11).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6ea1ae53dc4fc7f0e98cd8e5316cead2ddbe46b2 [formerly a97935dab4fe12777c1e79d8f2d4cbb8d53a36fb] [formerly a97935dab4fe12777c1e79d8f2d4cbb8d53a36fb [formerly cf424f5fd89bb5cd09b24c96633f8951b6fd7b54]]
Former-commit-id: 13de2b1be08bc07c99bf52496e915b1690a096e2
Former-commit-id: 6f6b7e25ae07f01492013c9aeba760afa82ac0c8