t0008: skip trailing space test on Windows
The Windows API does not preserve file names with trailing spaces (and
dots), but rather strips them. Our tools (MSYS bash, git) base the POSIX
emulation on the Windows API. As a consequence, it is impossible for bash
on Windows to allocate a file whose name has trailing spaces, and for git
to stat such a file. Both operate on a file whose name has the spaces
stripped. Skip the test that needs such a file name.

Note that we do not use (another incarnation of) prerequisite FUNNYNAMES.
The reason is that FUNNYNAMES is intended to represent a property of the
file system. But the inability to have trailing spaces in a file name is
a property of the Windows API. The file system (NTFS) does not have this
limitation.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: da98c6d2aa455f378dcb5ce0a3fb24b86ba527ad [formerly da98c6d2aa455f378dcb5ce0a3fb24b86ba527ad [formerly 35e4d775875603c748a127c82909f5d62a410ac8]]
Former-commit-id: 55a3e3c4fb4f5e78b3f2a301bc5551e4e8e44d85
Former-commit-id: 0de75ce6a15d474d3ba93acb9c8fa6a30eb68f06