t7810: add missing variables to tests in loop
Some tests in t7810-grep.sh are in a loop that runs them against HEAD and
the work tree.  In order for that to work the test code should use the
variables $L (display name), $H (HEAD or empty string) and $HC (revision
prefix for result lines); otherwise tests are just repeated with the same
target.  Add the variables where they're missing and make sure the test
description is wrapped in double quotes (instead of single quotes) to
allow variables to be expanded.

Signed-off-by: Rene Scharfe <l.s.r@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f04615c07eda261917919121eeae9ad4b0075806 [formerly f04615c07eda261917919121eeae9ad4b0075806 [formerly 9afad7a1e6363ed249a5cfcc95a35900ae4970cb]]
Former-commit-id: 959bd773ec14fb51c9e5ae85ca02160c0c730dd4
Former-commit-id: 3f17ea04e14d37e77750de490fd317044ab7d538