request-pull: documentation updates
The original description talked only about what it does.  Instead,
start it with the purpose of the command, i.e. what it is used for,
and then mention what it does to achieve that goal.

Clarify what <start>, <url> and <end> means in the context of the
overall purpose of the command.

Describe the extended syntax of <end> parameter that is used when
the local branch name is different from the branch name at the
repository the changes are published.

Helped-by: Eric Sunshine
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3c8e2387204c32ea0b22968a44d1c540249621d4 [formerly 3c8e2387204c32ea0b22968a44d1c540249621d4 [formerly ec445074e0ac4d6003acd21c512c43c8fdb8be86]]
Former-commit-id: ef5d129b162a6a120de43d4eacf0f15e4db07eb3
Former-commit-id: dae59c1c3dbf2df623f4fb5cf4b9159e5ad3e310