cat-file: restore warn_on_object_refname_ambiguity flag
Commit 25fba78 turned off the object/refname ambiguity check
during "git cat-file --batch" operations. However, this is a
global flag, so let's restore it when we are done.

This shouldn't make any practical difference, as cat-file
exits immediately afterwards, but is good code hygeine and
would prevent an unnecessary surprise if somebody starts to
call cmd_cat_file later.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ae40b1cffd5b74fc778cfd41f33a86a647846854 [formerly ae40b1cffd5b74fc778cfd41f33a86a647846854 [formerly a42fcd15d82b6660377749064e0ad25742ccc6f7]]
Former-commit-id: 09354994ed8d4c96ed202baf2aab4183d347d24e
Former-commit-id: 21e2f3d48bfbae0418166f632a7016cfd2f0cfd3