wt-status: make full label string to be subject to l10n
Earlier in 3651e45c (wt-status: take the alignment burden off
translators, 2013-11-05), we assumed that it is OK to make the
string before the colon in a label string we give as the section
header of various kinds of changes (e.g. "new file:") translatable.

This assumption apparently does not hold for some languages,
e.g. ones that want to have spaces around the colon.

Also introduce a static label_width to avoid having to run
strlen(padding) over and over.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 83ea3b28c98cc9c58febe1e9435e11918a469a5c [formerly 501f26154a84a75fdfe0c18ee8adb1c57478ea31] [formerly 501f26154a84a75fdfe0c18ee8adb1c57478ea31 [formerly d52cb5761a0e92e6051ea92edc7ae96fb6dca78f]]
Former-commit-id: ad9657ef66655c79f938f0b3cd5b08f943dec6eb
Former-commit-id: a93cfdc0cf6a783fc2e2863d253d18efa484fd70