rebase -i: do not "echo" random user-supplied strings
In some places we "echo" a string that comes from a commit log
message, which may have a backslash sequence that is interpreted by
the command (POSIX.1 allows this), most notably "dash"'s built-in
'echo'.

A commit message which contains the string '\n' (or ends with the
string '\c') may result in a garbage line in the todo list of an
interactive rebase which causes the rebase to fail.

To reproduce the behavior (with dash as /bin/sh):

  mkdir test && cd test && git init
  echo 1 >foo && git add foo
  git commit -m"this commit message ends with '\n'"
  echo 2 >foo && git commit -a --fixup HEAD
  git rebase -i --autosquash --root

Now the editor opens with garbage in line 3 which has to be
removed or the rebase fails.

Signed-off-by: Uwe Storbeck <uwe@ibr.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 78ecb41334c880619d43af5ad8c721a13032d905 [formerly 78ecb41334c880619d43af5ad8c721a13032d905 [formerly 47be06602656ee9cac860f675d2c8d1f0deabdbe]]
Former-commit-id: bcdbfde26795cb5bc98fcfb411fc1c6b82bff5d1
Former-commit-id: 79190b0c9eb6207d91a4b6fa13f677c9cc6a1aa7