pack-objects: show reused packfile objects in "Counting objects"
When we are sending a pack for push or fetch, we may reuse a
chunk of packfile without even parsing it. The progress
meter then looks like this:

  Reusing existing pack: 3440489, done.
  Counting objects: 3, done.

The first line shows that we are reusing a large chunk of
objects, and then we further count any objects not included
in the reused portion with an actual traversal.

These are all implementation details that the user does not
need to care about. Instead, we can show the reused objects
in the normal "counting..." progress meter (which will
simply go much faster than normal), and then continue to add
to it as we traverse.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 24d48afa5098013f4742e72f3644d5c7068d7699 [formerly 24903c01abe1eaf4904f52e44349d50c574f5023] [formerly 24903c01abe1eaf4904f52e44349d50c574f5023 [formerly 78d2214eb4d10ea1e30dd7e69a4e6d73d9f66164]]
Former-commit-id: 9a4bd3064e23ea7ce2b2d7bbb4bf5eb218e09115
Former-commit-id: 56d66a2f1af31a8ac85a9156f4c49144f58ddc63