shallow: verify shallow file after taking lock
Before writing the shallow file, we stat() the existing file
to make sure it has not been updated since our operation
began. However, we do not do so under a lock, so there is a
possible race:

  1. Process A takes the lock.

  2. Process B calls check_shallow_file_for_update and finds
     no update.

  3. Process A commits the lockfile.

  4. Process B takes the lock, then overwrite's process A's
     changes.

We can fix this by doing our check while we hold the lock.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 90ccf25b53716a0173efdcddc914622e2dc63d0c [formerly a5aab2ef924f03894aa19f613e3f32e496c28d94] [formerly a5aab2ef924f03894aa19f613e3f32e496c28d94 [formerly 7839632167bc6ceef20f28bd046f7001493b070f]]
Former-commit-id: 2f07f7e69d6ceda1592f956c0a91e3b7b0567166
Former-commit-id: d82cbe36d936449e947fe89636f77540de88e869