tests: use "env" to run commands with temporary env-var settings
Ordinarily, we would say "VAR=VAL command" to execute a tested
command with environment variable(s) set only for that command.
This however does not work if 'command' is a shell function (most
notably 'test_must_fail'); the result of the assignment is retained
and affects later commands.

To avoid this, we used to assign and export environment variables
and run such a test in a subshell, like so:

        (
                VAR=VAL && export VAR &&
                test_must_fail git command to be tested
        )

But with "env" utility, we should be able to say:

        test_must_fail env VAR=VAL git command to be tested

which is much shorter and easier to read.

Signed-off-by: David Tran <unsignedzero@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a2ca605e0bcf199a66bf6024eac98dc8bced7684 [formerly a2ca605e0bcf199a66bf6024eac98dc8bced7684 [formerly 512477b175288a429aaf4071dc9fe94f17f0e4ee]]
Former-commit-id: 81783ce43594738c5547967abf5954a231da1937
Former-commit-id: 3e384955f946ac8ca1c609a957ee5c123ea71570