diff-no-index: rename read_directory()
In the next patch, we will replace a manual checking of "." or ".."
with a call to is_dot_or_dotdot() defined in dir.h.  The private
function read_directory() defined in this file will conflict with
the global function declared there when we do so.

As a preparatory step, rename the private read_directory() to avoid
the name collision.

Helped-by: Eric Sunshine <sunshine@sunshineco.com>
Signed-off-by: Brian Bourn <ba.bourn@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c6f4566a451065da49bbcdc038f286fa6dfb13e1 [formerly c6f4566a451065da49bbcdc038f286fa6dfb13e1 [formerly 9daf0ef065bfef4fb9dcc1710a39a8177695c84f]]
Former-commit-id: 3288f8fb76a2a5cec908e832376dda8941a02929
Former-commit-id: 21a1c03e94bdae3128ad70a5afa344562283f696