t5510: Do not use $(pwd) when fetching / pushing / pulling via rsync
On MINGW, "pwd" is defined as "pwd -W" in test-lib.sh. This usually is the
right thing, but the absolute Windows path with a colon confuses rsync. We
could use $PWD in this case to work around the issue, but in fact there is
no need to use an absolute path in the first place, so get rid of it.

This was discovered in the context of the mingwGitDevEnv project and only
did not surface before with msysgit because the latter does not ship
rsync.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>
Acked-by: Johannes Schindelin <Johannes.Schindelin@gmx.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5d747bdc53dc321a5ca277ff9c76a8dfdad8e804 [formerly 5d747bdc53dc321a5ca277ff9c76a8dfdad8e804 [formerly ab58289eff9d305a8f911c8b7bd5b4e84a693563]]
Former-commit-id: b25b639a44fac52a8d7e60de35270c824d06230b
Former-commit-id: e812b1c7a23378d41cd4ef06a4be3a97acce9982