fsck: use bitwise-or assignment operator to set flag
fsck_tree() has two different ways to set a flag variable, either by
using a if-statement that guards an assignment, or by using a
bitwise-or assignment operator.  Most are done with the former, and
only one variable is assigned with the latter.

Since all the conditions are short-and-sweet, we can afford to
uniformly use the latter style, which makes the resulting code
shorter and easier to read.

Signed-off-by: Hiroyuki Sano <sh19910711@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5d7023059764c4535559f1788db4fb0654a9b439 [formerly 5d7023059764c4535559f1788db4fb0654a9b439 [formerly effd12ec876df016d4346fee0d3d6bf31308fa11]]
Former-commit-id: eef4d53adb4e1e6ab5188f55d2e226a410eef70a
Former-commit-id: 9593bf5d11ded7a5995c5653aebad3916f40c7cf