t: drop useless sane_unset GIT_* calls
Several test scripts manually unset GIT_CONFIG and other
GIT_* variables. These are generally taken care of for us by
test-lib.sh already.

Unsetting these is not only useless, but can be confusing to
a reader, who may wonder why some tests in a script unset
them and others do not (t0001 is particularly guilty of this
inconsistency, probably because many of its tests predate
the test-lib.sh environment-cleansing).

Note that we cannot always get rid of such unsetting. For
example, t9130 can drop the GIT_CONFIG unset, but not the
GIT_DIR one, because lib-git-svn.sh sets the latter. And in
t1000, we unset GIT_TEMPLATE_DIR, which is explicitly
initialized by test-lib.sh.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 09b6fb06ba0d35f6dbbfbc2e8ba9d1257de13c6b [formerly 09b6fb06ba0d35f6dbbfbc2e8ba9d1257de13c6b [formerly 221bf985063737240124fe15c328cd125e9d048e]]
Former-commit-id: 05c7005937e8f6ab3a2a8618d81607b073818c06
Former-commit-id: b946bed1090877174740f070a1d4d917fa6e5a38