t: stop using GIT_CONFIG to cross repo boundaries
Some tests want to check or set config in another
repository. E.g., t1000 creates repositories and makes sure
that their core.bare and core.worktree settings are what we
expect. We can do this with:

  GIT_CONFIG=$repo/.git/config git config ...

but it better shows the intent to just enter the repository
and let "git config" do the normal lookups:

  (cd $repo && git config ...)

In theory, this would cause us to use an extra subshell, but
in all such cases, we are actually already in a subshell.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4a3a76f9f37925f7735eab32336527a227a769bc [formerly 4a3a76f9f37925f7735eab32336527a227a769bc [formerly 3cc6a6f0f7519733deeebe17d7a19a53e9c02028]]
Former-commit-id: b658708f6e8d2319e46ac680e7723fcd3f8bfe99
Former-commit-id: 1750b07528633b5ff3b5a1c2290d0e752287b582