t: prefer "git config --file" to GIT_CONFIG
Doing:

  GIT_CONFIG=foo git config ...

is equivalent to:

  git config --file=foo ...

The latter is easier to read and slightly less error-prone,
because of issues with one-shot variables and shell
functions (e.g., you cannot use the former with
test_must_fail).

Note that we explicitly leave one case in t1300 which checks
the same operation on both GIT_CONFIG and "git config
--file". They are equivalent in the code these days, but
this will make sure it remains so.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fa5efd41f54bcfc09fc571c28aae2c3be585319b [formerly fa5efd41f54bcfc09fc571c28aae2c3be585319b [formerly f7e87141011fc55f8267ea76d3cc016ee87f9053]]
Former-commit-id: c7bbbd15eed1056155035ffe7a29e454f3b72240
Former-commit-id: eb7377ab965e4c655059cfca0593629975ce1db0