t0001: make symlink reinit test more careful
In the final test of t0001, we have a repo whose .git is a
symlink to a directory "here", and we use
"--separate-git-dir" to migrate that to a .git file pointing
to a different directory. We check that the data is migrated
to the new directory and that .git looks like a git-file.

We also check that "here" is not a directory, which is
slightly misleading. It should not be a directory, but
neither should it be gone. It is the actual resting place of
the git-file, and .git remains a symlink to it.

Let's check that more explicitly, both to make our test more
robust, and to make further cleanups in this area more
obvious.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7b19bdab5c23ff91df86dbcfbd9a3bd64e5d52c7 [formerly 7b19bdab5c23ff91df86dbcfbd9a3bd64e5d52c7 [formerly 3d06c5f19daf6710d99b6e4c7c0d097c8ce528fc]]
Former-commit-id: d4c6c88b0b7e3c9a948468cc5c52589a321b5d98
Former-commit-id: 9b13544a66b3da46550b8608f3a5b785c1dc0da9