t0001: use test_must_fail
We've hand-rolled several "if" statements looking for
failures. We can use test_must_fail here, which is shorter
and more robust.

Note that we modify the commands slightly (to use "git init
foo" rather than "cd foo && git init") to avoid dealing with
a subshell, but this should not affect the outcome.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 32be17e21fa3e4356e68d0ae87be5f0888ecefa4 [formerly 32be17e21fa3e4356e68d0ae87be5f0888ecefa4 [formerly 0981140fcc2fe33c37da87802b95e01af2a94add]]
Former-commit-id: 279a9105679b1e16afd92299cec89d7dda8d8300
Former-commit-id: 62160ec3abf87ee0878cb524097ce9dce64d6fcc