remote-hg: do not fail on invalid bookmarks
Mercurial can have bookmarks pointing to "nullid" (the empty root
revision), while Git can not have references to it. When cloning or
fetching from a Mercurial repository that has such a bookmark, the
import failed because git-remote-hg was not be able to create the
corresponding reference.

Warn the user about the invalid reference, and do not advertise these
bookmarks as head refs, but otherwise continue the import. In
particular, we still keep track of the fact that the remote repository
has a bookmark of the given name, in case the user wants to modify that
bookmark.

Also add some test cases for this issue.

Reported-by: Antoine Pelisse <apelisse@gmail.com>
Signed-off-by: Max Horn <max@quendi.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 14c4f9e2d3a93fa3bdd316b14d93f0bf279f0555 [formerly 14c4f9e2d3a93fa3bdd316b14d93f0bf279f0555 [formerly 51be46ec4d0b64c1deb60a4814bcd24b6b478eeb]]
Former-commit-id: 30ce2c8b841146d9c1ac38cf615b4e4cf4443ab4
Former-commit-id: 738a61607a1bc5a8ee3b8f6753479eeae724b96e