t4018: reduce test files for pattern compilation tests
All test cases that need a file with specific text patterns have been
converted to utilize texts in the t4018/ directory. The remaining tests
in the test script deal only with the validity of the regular
expressions. These tests do not depend on the contents of files that
'git diff' is invoked on. Remove the largish here-document and use only
tiny files.

While we are touching these tests, convert grep to test_i18ngrep as the
texts checked for may undergo translation in the future.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b44d30cb46bc311442cef2da25386a22d69b21b5 [formerly b44d30cb46bc311442cef2da25386a22d69b21b5 [formerly ad5070fb363b1e7ee7e1df83a17f4e8e1eb607a0]]
Former-commit-id: 5c9ee58df2888561465a4e8d6cbf087fdd30033b
Former-commit-id: 6c5edeac5fb56f6dbf3abd5769fd94e6f9ff0023