t1502: protect runs of SPs used in the indentation
The expected output from the argument help use runs of SPs to align
the description of each option; a careless use of --whitespace=fix
can turn leading parts of them into appropriate number of HTs.
Prevent such a breakage by prefixing all the expected lines with
leading vertical bars in the original and stripping them with a
small sed script.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 18aa30c61deba47b9bcd6b13855c39b75a6020fd [formerly 18aa30c61deba47b9bcd6b13855c39b75a6020fd [formerly ce7f8745aa773dfe2fc0a1b8c327b50c7ce4a31e]]
Former-commit-id: c6b178b8de19f10a5b487759ecba9fd32aaaef59
Former-commit-id: 5bc0e711bf942e7dff2d3d25b4336576ecfc42da