parse-options: multi-word argh should use dash to separate words
"When you need to use space, use dash" is a strange way to say that
you must not use a space.  Because it is more common for the command
line descriptions to use dashed-multi-words, you do not even want to
use spaces in these places.  Rephrase the documentation to avoid
this strangeness.

Fix a few existing multi-word argument help strings, i.e.

 - GPG key-ids given to -S/--gpg-sign are "key-id";
 - Refs used for storing notes are "notes-ref"; and
 - Expiry timestamps given to --expire are "expiry-date".

and update the corresponding documentation pages.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b6e8f68ee0275f70fb95d7317aaa76d50fe93f47 [formerly b6e8f68ee0275f70fb95d7317aaa76d50fe93f47 [formerly e703d7118c68bb5b1f850dae4060609a88500b18]]
Former-commit-id: 2ca8186a624b98a6c298bb9b3c017ac3d6bae745
Former-commit-id: 0ea0c8f9007a17ba2b0905296316c045db47f42c