check-builtins.sh: use the $(...) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

    for _f in $(find . -name "*.sh")
    do
      sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
    done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 984861537e5fb705092ac16ecb6a2bb130f9c630 [formerly 984861537e5fb705092ac16ecb6a2bb130f9c630 [formerly b09d8552bd04b17c20f89b73e09206719f9e7b84]]
Former-commit-id: 0a27f32266d8bbc148307c40f9d320944790ae80
Former-commit-id: 1f6491e1883a15dfed1bb6cc9c381d2e746fe2fa