git-am.sh: use the $(...) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

    for _f in $(find . -name "*.sh")
    do
      sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
    done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 60f071bf725a70e3397d7f050ce92c9de6c6a323 [formerly 60f071bf725a70e3397d7f050ce92c9de6c6a323 [formerly 75ee3d7078fbcc3b87a3ae5e0cf42f46256c5da4]]
Former-commit-id: a627ca0b527e49119eabd5ca34c25e8ef1aa9225
Former-commit-id: 1c37a2a639173560b941939d0874a7fc2dca0349