tests: don't rely on strerror text when testing rmdir failure
AIX doesn't make a distiction between EEXIST and ENOTEMPTY; relying
on the strerror string for the rmdir failure is fragile. Just test
that the start of the string matches the Git controlled "failed to
rmdir..."  error. The exact text of the OS generated error string
isn't important to the test.

Signed-off-by: Charles Bailey <cbailey32@bloomberg.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5e23cc615673343876263f60aafb0871a44ca184 [formerly 5e23cc615673343876263f60aafb0871a44ca184 [formerly 11b53902513c6d59f59594845ce53fe761f0894c]]
Former-commit-id: e2ddb20feda363a0dab3a40eea7bc5169e663cf6
Former-commit-id: 55ee39d924aa9a3d953c79dbcead2710f18c5d3c