t4212: loosen far-in-future test for AIX
One of the tests in t4212 checks our behavior when we feed
gmtime a date so far in the future that it gives up and
returns NULL. Some implementations, like AIX, may actually
just provide us a bogus result instead.

It's not worth it for us to come up with heuristics that
guess whether the return value is sensible or not. On good
platforms where gmtime reports the problem to us with NULL,
we will print the epoch value. On bad platforms, we will
print garbage.  But our test should be written for the
lowest common denominator so that it passes everywhere.

Reported-by: Charles Bailey <cbailey32@bloomberg.net>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d2a5da7619c38ca63530a51794dd28fc7fd0bfe9 [formerly d2a5da7619c38ca63530a51794dd28fc7fd0bfe9 [formerly f80d1f95f0adf9909f42eb9e74546963daa4b0c1]]
Former-commit-id: ab4a0de72d1dae245dca5b10ce3aae2d84da49e3
Former-commit-id: 5d0d569af0eb4a4d4bc93dc3e38322c6b2bb154a