date: recognize bogus FreeBSD gmtime output
Most gmtime implementations return a NULL value when they
encounter an error (and this behavior is specified by ANSI C
and POSIX).  FreeBSD's implementation, however, will simply
leave the "struct tm" untouched.  Let's also recognize this
and convert it to a NULL (with this patch, t4212 should pass
on FreeBSD).

Reported-by: René Scharfe <l.s.r@web.de>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7274d149aa6bbbc77572da98f113a02c20a0cdf3 [formerly 2eea3bbc63dc716b2f3cf4aa2b489143648f291b] [formerly 2eea3bbc63dc716b2f3cf4aa2b489143648f291b [formerly 6654754779d2a90af91a9c6b93d7e4e7ee16cfab]]
Former-commit-id: 7764bd5f5f9c2b5a1f36820ea1649c0d3701967f
Former-commit-id: 6c3149b96c2e637398b72e2996b270f336f1cdd8