t1400: provide more usual input to the command
The old version was passing (among other things)

    update SP refs/heads/c NUL NUL 0{40} NUL

to "git update-ref -z --stdin" to test whether the old-value check for
c is working.  But the <newvalue> is empty, which is a bit off the
beaten track.

So, to be sure that we are testing what we want to test, provide an
actual <newvalue> on the "update" line.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: aa76a6f52554e460c0cd6d13661ec29ad45249c8 [formerly aa76a6f52554e460c0cd6d13661ec29ad45249c8 [formerly c13291108880f9adc26a2ab5c09535869afecb22]]
Former-commit-id: 37c2b398106051c668d8ddde0fc391f2749af1f3
Former-commit-id: 1cf716122dc941a64b842c0bc090c09866865dca