update_refs(): fix constness
The old signature of update_refs() required a
(const struct ref_update **) for its updates_orig argument.  The
"const" is presumably there to promise that the function will not
modify the contents of the structures.

But this declaration does not permit the function to be called with a
(struct ref_update **), which is perfectly legitimate.  C's type
system is not powerful enough to express what we'd like.  So remove
the first "const" from the declaration.

On the other hand, the function *can* promise not to modify the
pointers within the array that is passed to it without inconveniencing
its callers.  So add a "const" that has that effect, making the final
declaration
(struct ref_update * const *).

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3fa24a873b289bb4220f853e0caa3f7943aef063 [formerly 3fa24a873b289bb4220f853e0caa3f7943aef063 [formerly 595deb8da69b4f816ff0c8e669b49f7527ff609b]]
Former-commit-id: 92158be58aca2921d8eb86bd25eaa04c37d21d82
Former-commit-id: 5a302c016031e29063d5c41da4fdcfb354b29883