update-ref --stdin: read the whole input at once
Read the whole input into a strbuf at once, and then parse it from
there.  This might also be a tad faster, but that is not the point.
The point is to decouple the parsing code from the input source (the
old parsing code had to read new data even in the middle of commands).
Add docstrings for the parsing functions.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 00198eef9dfcba4ba2a187d1150596bc8de6b4db [formerly 00198eef9dfcba4ba2a187d1150596bc8de6b4db [formerly e23d84350a3f4a3bfb86037eb1e6c4b28240324e]]
Former-commit-id: 82f42a6d4b20df123d2c96059955bf6f13c376b9
Former-commit-id: ecaebb385d23e37c33a85f8a48bb40619c971764