update-ref.c: extract a new function, parse_next_sha1()
Replace three functions, update_store_new_sha1(),
update_store_old_sha1(), and parse_next_arg(), with a single function,
parse_next_sha1().  The new function takes care of a whole argument,
including checking whether it is there, converting it to an SHA-1, and
emitting errors on EOF or for invalid values.  The return value
indicates whether the argument was present or absent, which requires
a bit of intelligence because absent values are represented
differently depending on whether "-z" was used.

The new interface means that the calling functions, parse_cmd_*(),
don't have to interpret the result differently based on the
line_termination mode that is in effect.  It also means that
parse_cmd_create() can distinguish unambiguously between an empty new
value and a zeros new value, which fixes a failure in t1400.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 03aa418961f8f79c1c033cd83fd4a2c8b7b869e5 [formerly 03aa418961f8f79c1c033cd83fd4a2c8b7b869e5 [formerly 3afcc4637452100c68b469de7757dd2b45b4d29c]]
Former-commit-id: 7bf4376191992c2447e0060d5a0e59b1282ac741
Former-commit-id: 82671bf2c8c1422fa16d5373144de5ae3d7d1ba6