update-ref --stdin -z: deprecate interpreting the empty string as zeros
In the original version of this command, for the single case of the
"update" command's <newvalue>, the empty string was interpreted as
being equivalent to 40 "0"s.  This shorthand is unnecessary (binary
input will usually be generated programmatically anyway), and it
complicates the parser and the documentation.

So gently deprecate this usage: remove its description from the
documentation and emit a warning if it is found.  But for reasons of
backwards compatibility, continue to accept it.

Helped-by: Brad King <brad.king@kitware.com>
Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8a2761f4b9da8ff964b6b99be216df8d96a1d02d [formerly 8a2761f4b9da8ff964b6b99be216df8d96a1d02d [formerly 1fbd504942b20a541ba4fcbe90d3ea21b03717e4]]
Former-commit-id: ca4bc666e4cb754ae6f0242d632792b755dedeaf
Former-commit-id: 57cf48a6ea466a15673a901a5ef5484c6d5b5c88