refs: add a concept of a reference transaction
Build out the API for dealing with a bunch of reference checks and
changes within a transaction.  Define an opaque ref_transaction type
that is managed entirely within refs.c.  Introduce functions for
beginning a transaction, adding updates to a transaction, and
committing/rolling back a transaction.

This API will soon replace update_refs().

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d2d1c2c76091486e9f7c2099090261f6168fc390 [formerly d2d1c2c76091486e9f7c2099090261f6168fc390 [formerly caa4046c4f480ceae5afb20e3172a437865cc51f]]
Former-commit-id: 4f90c37f70c00f816574f3a3aee184c3cf87a181
Former-commit-id: a2ad0c36290d657f641a16b9bc7b8ef2e2db5732