git-gui: show staged submodules regardless of ignore config
Currently setting submodule.<name>.ignore and/or diff.ignoreSubmodules to
"all" suppresses all output of submodule changes for git-gui. This is
really confusing, as even when the user chooses to record a new commit for
an ignored submodule by adding it manually this change won't show up under
"Staged Changes (Will Commit)".

Fix that by using the '--ignore-submodules=dirty' option for both callers
of "git diff-index --cached" when the underlying git version supports that
option.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>

Former-commit-id: caad94511e1f62b95ae62358491d3817a296e9b6 [formerly 5ee578e1f2da10092059f533dde4949aa54bc517] [formerly 5ee578e1f2da10092059f533dde4949aa54bc517 [formerly e0db1dd7d4bc87825f0d7d743f1a01b4461a59ff]]
Former-commit-id: f7a485c8e5c929269acce17744a1503608a22bd8
Former-commit-id: 162fea4d009610dc121892e4ff3a291718112972