gitk: Show staged submodules regardless of ignore config
Currently setting submodule.<name>.ignore and/or diff.ignoreSubmodules to
"all" suppresses all output of submodule changes for gitk. This is really
confusing, as even when the user chooses to record a new commit for an
ignored submodule by adding it manually this change won't show up under
"Local changes checked in to index but not committed".

Fix that by using the '--ignore-submodules=dirty' option for both callers
of "git diff-index --cached" when the underlying git version supports that
option.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>

Former-commit-id: ee6fef055584e78cea9b6575889fced947821b2c [formerly bc43c28d16d22c703b67d712cfddfce203ec5152] [formerly bc43c28d16d22c703b67d712cfddfce203ec5152 [formerly 17f9836c8afc8d4b6d6bb91fd0658df090d75c8f]]
Former-commit-id: 6acf8515f2c59252b91a4dd139d4b0e209416c90
Former-commit-id: 50426b43f2215cf8a47bf67ad9e8aea3eb7f5a09