completion: fix completing args of aliased "push", "fetch", etc.
Some commands need the first word to determine the actual action that is
being executed, however, the command is wrong when we use an alias, for
example 'alias.p=push', if we try to complete 'git p origin ', the
result would be wrong because __git_complete_remote_or_refspec() doesn't
know where it came from.

So let's override words[1], so the alias 'p' is override by the actual
command, 'push'.

Reported-by: Aymeric Beaumet <aymeric.beaumet@gmail.com>
Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b13d748f34cfd2c07abcf582d59ab18ed3d25e7c [formerly b13d748f34cfd2c07abcf582d59ab18ed3d25e7c [formerly 880111c11befaaf070e8ea1021bf11b4fb58c461]]
Former-commit-id: 0c4d4fcd79161d93623296efe2511c2cc3dd4c71
Former-commit-id: 25f2092daac7e44128ec9d0fc0dbc3dc1763dbff