wrapper.c: add xpread() similar to xread()
It is a common mistake to call read(2)/pread(2) and forget to
anticipate that they may return error with EAGAIN/EINTR when the
system call is interrupted.

We have xread() helper to relieve callers of read(2) from having to
worry about it; add xpread() helper to do the same for pread(2).

Update the caller in the builtin/index-pack.c and the mmap emulation
in compat/.

Signed-off-by: Yiannis Marangos <yiannis.marangos@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: cfc657e8411ddd18b26fd5740bba3a6c05c85439 [formerly 8160b4240e9ad5c3449878d22beadce22f1c4894] [formerly 8160b4240e9ad5c3449878d22beadce22f1c4894 [formerly 9aa91af0361e3c32fde5f8388dee963838308cd6]]
Former-commit-id: 4956f6937ad5721d879d9a65a4617db9d7ea24fc
Former-commit-id: 0d9aa854d654f16903c853ec6b732ea3777ce939