test: fix t7001 cp to use POSIX options
Since 11502468 and 04c1ee57 (both first appearing in v1.8.5), the
t7001-mv test has used "cp -a" to perform a copy in several of the
tests.

However, the "-a" option is not required for a POSIX cp utility and
some platforms' cp utilities do not support it.

The POSIX equivalent of -a is -R -P -p.

Change "cp -a" to "cp -R -P -p" so that the t7001-mv test works
on systems with a cp utility that only implements the POSIX
required set of options and not the "-a" option.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e31728e2b7bf942f6cb4b0f0428bed9221eb7d80 [formerly e31728e2b7bf942f6cb4b0f0428bed9221eb7d80 [formerly 00764ca10e0b55b21853fe3f41d688254cfaec08]]
Former-commit-id: 8392fc72da3975be5146bb3c11de68a47873b88b
Former-commit-id: b6658361a07cdad302e98a12cf05753b8cdc6406