transport-helper: fix sync issue on crashes
When a remote helper crashes while pushing we should revert back to the
state before the push, however, it's possible that `git fast-export`
already finished its job, and therefore has exported the marks already.

This creates a synchronization problem because from that moment on
`git fast-{import,export}` will have marks that the remote helper is not
aware of and all further commands fail (if those marks are referenced).

The fix is to tell `git fast-export` to export to a temporary file, and
only after the remote helper has finishes successfully, move to the
final destination.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1ef6440e9047ab4a82d9026a753b22f4e6b0bd11 [formerly 1ef6440e9047ab4a82d9026a753b22f4e6b0bd11 [formerly 3994e64d77d644df0f290f8bf5210ff0cb0adde2]]
Former-commit-id: 364f474e39685a93d1ee48d1f79862009650b300
Former-commit-id: c7f11fb9063be6f66859937578c05ac0e2479f4b