Makefile: use curl-config to determine curl flags
curl-config should always be installed alongside a curl distribution,
and its purpose is to provide flags for building against libcurl, so
use it instead of guessing flags and dependent libraries.

Allow overriding CURL_CONFIG to a custom path to curl-config, to
compile against a curl installation other than the first in PATH.

Depending on the set of features curl is compiled with, there may be
more libraries required than the previous two options of -lssl and
-lidn. For example, with a vanilla build of libcurl-7.36.0 on Mac OS X
10.9:

$ ~/d/curl-out-7.36.0/lib/curl-config --libs
-L/Users/dborowitz/d/curl-out-7.36.0/lib -lcurl -lgssapi_krb5 -lresolv -lldap -lz

Use this only when CURLDIR is not explicitly specified, to continue
supporting older builds.

Signed-off-by: Dave Borowitz <dborowitz@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a27de25737ca8343fd5554dc901e81cba4996816 [formerly 1e9797fc64a344c2acb27b9acc868b7e4d529e3a] [formerly 1e9797fc64a344c2acb27b9acc868b7e4d529e3a [formerly 61a64fff4f710f43826c8446d3fd8f3ae78c0411]]
Former-commit-id: 1f068f7b50c5b0d4d12801b6be573231048320c4
Former-commit-id: 7364b350e014a778f4814b975fc7bab0589bead9