howto-index.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 59f3f78e2e7d586c95aca8f77aec0f8102720b56 [formerly 59f3f78e2e7d586c95aca8f77aec0f8102720b56 [formerly f25f5e61a7c1a22dd85413d61db639e7c01268f4]]
Former-commit-id: 600d00cb924a7944f72b9ee2be696d71ad1fbd5c
Former-commit-id: aeb4fd3992b8c27bd04d1a1b3c217a5c83e58df6