install-webdoc.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 79cedb341e94645f48925ebfb28997c22bee8c59 [formerly 1c673cf6c4a1e924b3c72e0dc759b5d795908dd4] [formerly 1c673cf6c4a1e924b3c72e0dc759b5d795908dd4 [formerly 2c4a050bc60bf7a0633aa1460a8538d18df67848]]
Former-commit-id: b999cca4aa4921e20aa620d33c9b84234522641f
Former-commit-id: bd18bb06fe6a88e7aafcd539dfc36f8b40e972e3