git-checkout.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b7360a49a440f55221339231d7fb9cd0ed344739 [formerly 93a3096d7b96d1efd9e6cf7341902c8985878f09] [formerly 93a3096d7b96d1efd9e6cf7341902c8985878f09 [formerly 844cb24f28afd4febcfc1ddc44f9de47130acf4b]]
Former-commit-id: db6163f1f8a66797cbfbca13200048f00352f22a
Former-commit-id: 831afc3ff7744c5ee879345eafedbd32cd73d5be