git-clone.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b23e944cde1221977e5e1366778048fafcbaff48 [formerly 68c313214d0e667fff0aae21a380fc269743b3d0] [formerly 68c313214d0e667fff0aae21a380fc269743b3d0 [formerly add77e8400252ef59441770a96379dcad797b649]]
Former-commit-id: eaf95b6904560dda6d9723522c31c96bbea1f77f
Former-commit-id: da609232ad7f50f46f6e523bb1acea33bebc9180