git-commit.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c747125f269826066140d78c1851c0af6d021b8b [formerly c747125f269826066140d78c1851c0af6d021b8b [formerly 346b54dbc92bb3c928d326da9d8addf5eb50acce]]
Former-commit-id: 68733ed620b6e17ad336fb08a375a26644e39192
Former-commit-id: f741461a628e6581e8f5267477b9969e6da6373a