git-fetch.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9472be9d28d6493a90b4b3a46b99d3411abeebdb [formerly 9472be9d28d6493a90b4b3a46b99d3411abeebdb [formerly 3e867415170ed73511291fb2dd47bac3c4d51f3a]]
Former-commit-id: 91da3ed6521ab30aa9104af15b3174cc6c0d32a1
Former-commit-id: c16d9af1ae7a1b7a4b1b7429645dfbb210fe79e0