git-ls-remote.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a70f5861e6f95fa2efb8e07aadd89cdac0ee353c [formerly a70f5861e6f95fa2efb8e07aadd89cdac0ee353c [formerly 1b3cddd28834ef7c32f2049054c9179c68a381f6]]
Former-commit-id: 8dfa86dfc512ea6e6ae8d6fab3873a1eb8f025f9
Former-commit-id: dd7bbff670c778fc48b8b18cd282c73f5495dfcc