git-merge.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fe6a5f3096e4ba728d8701848bed6ee808453cea [formerly fe6a5f3096e4ba728d8701848bed6ee808453cea [formerly 34da37cc4230affae93483f6c51fca6beb6e6c64]]
Former-commit-id: aec49072f2f5c135a854acf9ee1442381a9686cc
Former-commit-id: 29763849027e810329565c90db5e203a38eebaf3