git-repack.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1299355fc363f198cebd8668de9ddb1a5aaa9b87 [formerly 1299355fc363f198cebd8668de9ddb1a5aaa9b87 [formerly ddbac79de939d5a85860f6cc250c239fbfff439a]]
Former-commit-id: 81f9f3b2f28aad506443e5a6bbba50e8d6b8596c
Former-commit-id: 66242eca3c891e263591ebf64eb3710afa9f348c