git-revert.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a17eadb2f35b433d57c31a7fb986b001afd25da2 [formerly a17eadb2f35b433d57c31a7fb986b001afd25da2 [formerly fb6644a32fa06ff2698ecc66f4008d2afdf0b9c7]]
Former-commit-id: 35ea30e5767392593eb4dd5fd67da092ffcaf3e8
Former-commit-id: 9a7249eb93c271e962c5a9c50e9132bf7cdc4559