git-tag.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8efc8dc71abec253cc516b690be6a65672119f59 [formerly 8efc8dc71abec253cc516b690be6a65672119f59 [formerly b352891021d74528f668306d0882633cc0673a89]]
Former-commit-id: bce6b71f53c3ad316a47e5e8ed5c47cec8309cdb
Former-commit-id: e2327de48ac45c3627993b960a8f888f3f111247