t9360-mw-to-git-clone.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 64d2efe12076d6c52cb407f7988847323eb704a5 [formerly 33c401d1384f9908613b36e704892827a5b61c95] [formerly 33c401d1384f9908613b36e704892827a5b61c95 [formerly c9b92706af6441ba91fec314b6aac7f105e61b22]]
Former-commit-id: 67bd351039874e0c0c7dbd019aaffdf8fc85a116
Former-commit-id: 8c52ac27bf15ff7066bdf64ae08c08a8a1354fd4