t9362-mw-to-git-utf8.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 015b73f296db606399728022b7e3fdfd53f28046 [formerly 9a538e3ac652008d3e774cbcd4b2bac1705bd117] [formerly 9a538e3ac652008d3e774cbcd4b2bac1705bd117 [formerly bd368a9baf5a42e2076ef4c8a35312ccab171452]]
Former-commit-id: 50c68e80ed3abc477940bc7b7b07765da5f7226d
Former-commit-id: 96455e1c63f626de5161fa8f57fe6c0578cdfa6f