commit.c: check for lock error and return early
Move the check for the lock failure to happen immediately after
lock_any_ref_for_update().  Previously the lock and the
check-if-lock-failed was separated by a handful of string
manipulation statements.

Moving the check to occur immediately after the failed lock makes
the code slightly easier to read and makes it follow the pattern of

 try-to-take-a-lock();
 if (check-if-lock-failed) {
    error();
 }

Signed-off-by: Ronnie Sahlberg <sahlberg@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4d1eb9da9f25afa505a6990cab27b48c98bd5fa5 [formerly 9dcddf175adb93ff8e6fe6cc118ab6580996459c] [formerly 9dcddf175adb93ff8e6fe6cc118ab6580996459c [formerly 55a5c8d72b03ad3727b89900069a1d58664fe5e4]]
Former-commit-id: 0db78a381a3462bcb617620559e867d05b5c6242
Former-commit-id: 1a2ab58e7159d78bc5ca8bb44170389b06733298