refs.c: change ref_transaction_create to do error checking and return status
Do basic error checking in ref_transaction_create() and make it return
non-zero on error. Update all callers to check the result of
ref_transaction_create(). There are currently no conditions in _create that
will return error but there will be in the future. Add an err argument that
will be updated on failure.

Signed-off-by: Ronnie Sahlberg <sahlberg@google.com>
Reviewed-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: aa9b06cab2a68413b4b578cb4efed4243b4e61d9 [formerly aa9b06cab2a68413b4b578cb4efed4243b4e61d9 [formerly b416af5bcdca3ac8426220f09efbfca2f1bec6e0]]
Former-commit-id: 93590c1eb366baad7b65731d40598134514a82d1
Former-commit-id: 3d33d16549a1dbc84c82834164ca1472c9575459