refs.c: update ref_transaction_delete to check for error and return status
Change ref_transaction_delete() to do basic error checking and return
non-zero on error. Update all callers to check the return for
ref_transaction_delete(). There are currently no conditions in _delete that
will return error but there will be in the future. Add an err argument that
will be updated on failure.

Signed-off-by: Ronnie Sahlberg <sahlberg@google.com>
Reviewed-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 927f6ac7324a6db4e7316c285883fbaa356863dd [formerly 927f6ac7324a6db4e7316c285883fbaa356863dd [formerly 8c8bdc0d3582e42c13815a191344b1aa3e06c792]]
Former-commit-id: ed22512088ed96f4ce7f28b5129dd38ac39b7dcf
Former-commit-id: 9e629532a4e58c4af8d9ca52c57e077d36fc38ab