walker.c: use ref transaction for ref updates
Switch to using ref transactions in walker_fetch(). As part of the refactoring
to use ref transactions we also fix a potential memory leak where in the
original code if write_ref_sha1() would fail we would end up returning from
the function without free()ing the msg string.

Note that this function is only called when fetching from a remote HTTP
repository onto the local (most of the time single-user) repository which
likely means that the type of collisions that the previous locking would
protect against and cause the fetch to fail for are even more rare.

Signed-off-by: Ronnie Sahlberg <sahlberg@google.com>
Reviewed-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6cdf1f96424496ed829fcd262fb1b9a629a422b9 [formerly 89605373091df6dd25b45f4afebfa1545c9f4809] [formerly 89605373091df6dd25b45f4afebfa1545c9f4809 [formerly b6b10bb44c31b5a830c1636c7eb8b122f9b67858]]
Former-commit-id: 9972811f4226c20fd1cab9e3f57f2caa7d59861b
Former-commit-id: 8981c33237590bc793199146014af00ee50d5b54