run_external_diff: use an argv_array for the command line
We currently generate the command-line for the external
command using a fixed-length array of size 10. But if there
is a rename, we actually need 11 elements (10 items, plus a
NULL), and end up writing a random NULL onto the stack.

Rather than bump the limit, let's just use an argv_array, which
makes this sort of error impossible.

Noticed-by: Max L <infthi.inbox@gmail.com>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4f53752df9fc089aab402638ca32b953af205306 [formerly 4f53752df9fc089aab402638ca32b953af205306 [formerly 82fbf269b9994d172719b2d456db5ef8453b323d]]
Former-commit-id: 57b6ee5733a6168c90c634fb24baef280aa56369
Former-commit-id: 7b62402137b42e4efd721065f0c86f332d43d914