t9117: use --prefix "" instead of --prefix=""
Versions of Perl's Getopt::Long module before 2.37 do not contain
this fix that first appeared in Getopt::Long version 2.37:

* Bugfix: With gnu_compat, --foo= will no longer trigger "Option
  requires an argument" but return the empty string.

Instead of using --prefix="" use --prefix "" when testing an
explictly empty prefix string in order to work with older versions
of Perl's Getopt::Long module.

Also add a paragraph on this workaround to the documentation of
git-svn itself.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>
Acked-by: Johan Herland <johan@herland.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 86f0db30f6d0d2b62c85960b2357a2d1d410faec [formerly d4dccabfedbabc6c10ac269c2101362f9e9c90f9] [formerly d4dccabfedbabc6c10ac269c2101362f9e9c90f9 [formerly 7bbc458b44873d0027261562327a20ebd6a97733]]
Former-commit-id: 9a948f47ea03bdc9e18fb8606be98afe12e9b330
Former-commit-id: 5c4ef417ec08046c9afe0a362431c5c4162751b3