blame: dynamic blame_date_width for different locales
When show date in relative date format for git-blame, the max display
width of datetime is set as the length of the string "Thu Oct 19
16:00:04 2006 -0700" (30 characters long).  But actually the max width
for C locale is only 22 (the length of string "x years, xx months ago").
And for other locale, it maybe smaller.  E.g. For Chinese locale, only
needs a half (16-character width).

Set blame_date_width as the display width of _("4 years, 11 months
ago"), so that translators can make the choice.

Helped-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Jiang Xin <worldhello.net@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3c738adfa52bcecb25dceaf04f60142c2b48351c [formerly 8b61cce6843ff4bb7bd6f3e958a64e774a9bda3e] [formerly 8b61cce6843ff4bb7bd6f3e958a64e774a9bda3e [formerly dd75553b357d4483d2b58a9e1ac86352f3d2e103]]
Former-commit-id: 66df7bf9a58948d0c8494f30bc1b8686629a00d9
Former-commit-id: 2cefd7e8e2d4470ac356fc22cfb5554b426eba65