t9365-continuing-queries.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2aae84b90b22c5ecf8d7b0e314aaa88fc6c3c5a0 [formerly 2aae84b90b22c5ecf8d7b0e314aaa88fc6c3c5a0 [formerly 9bfeaa0bcf2d7150836b583196e18f973c3be17f]]
Former-commit-id: 49256f4a9871abbd1b41fd6e7d0ccf5f43171a58
Former-commit-id: 00a44086376cf032e97b5d95bb48c360336a2f60