test-gitmw-lib.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: bfecbfca92136e2d045ee5561679a537364691d6 [formerly bfecbfca92136e2d045ee5561679a537364691d6 [formerly 0eca37c63a1e03e4f33366157eb406c8182bf0e9]]
Former-commit-id: 5fc0db4f07e76cf4f10056ceec2a6f15767050a1
Former-commit-id: 12050d76a207a0fd79a40c76fd2675b102bec2f8