t7900-subtree.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b409d6a212569f52cc2a59684b2f465d8bbfd981 [formerly b409d6a212569f52cc2a59684b2f465d8bbfd981 [formerly 5c00acdd25b0b214e511fa35faaff8595467c480]]
Former-commit-id: 2e62ebfd169e905a305b5eaad6b67d1f5e4bdd38
Former-commit-id: 0961749ff4c6d75b09c5f5465e04b04652d6e560