appp.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e53750940707d90a6847dac1b7af7eeab45c1743 [formerly e53750940707d90a6847dac1b7af7eeab45c1743 [formerly faf58f4ee6a5599c3b8790b770d6e1736174b7a3]]
Former-commit-id: 862d007c60eab704aef212d43a519a9361a7c7a8
Former-commit-id: 5cf0e0824cb983f64ca76200cecb0b06279d2c29