git-pull.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: de6f6998849afaad8e30276f1fb1ebb52e621725 [formerly de6f6998849afaad8e30276f1fb1ebb52e621725 [formerly eadf619cd434135bdfa598566376b3e3dbf9f2b5]]
Former-commit-id: 129ad442057e73b1d5211470414c9ef954842796
Former-commit-id: 3ab30810de2bf5349f4f23fc2f947f90d4c69e28