git-rebase--merge.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 12854b894442a6b7a5a46edca2796450fc3e97a5 [formerly 12854b894442a6b7a5a46edca2796450fc3e97a5 [formerly f257482c9ccbb27e70f190da6a8152469b0fed1b]]
Former-commit-id: e7fafc936a690ca9f9fbe20aae1610797770a059
Former-commit-id: c72306d550b2d6ad5d2fb9ed85c3f7fa4dfc6cb8