git-rebase.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: dfa7f5d54cdbf5fba62558d03d6ead1e140bfa4c [formerly dfa7f5d54cdbf5fba62558d03d6ead1e140bfa4c [formerly 728fc79c0016fd7133c32f1d193d0a243e552257]]
Former-commit-id: 7bd77bff637ffee7671573eb07bfda1257c82d3b
Former-commit-id: 32f0d74fd5c2776a948508c528f6c83ac85321e8