git-stash.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e472c7a12d6c834cf54ac9a877fc4d19496c5860 [formerly e472c7a12d6c834cf54ac9a877fc4d19496c5860 [formerly d0ea45bfc7129304016b24bdb18523b927d0ab17]]
Former-commit-id: 961097d2c8b0be61e44ef31b6d1cf3132994346f
Former-commit-id: b5451456c1482bfe1460fe4d956520fb7ed5676f