git-web--browse.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ca6755ec8fe7ad41cd0685a330b66651ad799855 [formerly ca6755ec8fe7ad41cd0685a330b66651ad799855 [formerly 9e5878fbede57c0499133adf73844261849cd7b2]]
Former-commit-id: 411612dd3060d3b86a2ba94753fe8ea751074376
Former-commit-id: 0afa0ba2e73bd7e315e3d9e22606147df68e2b18