lib-credential.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 29c62bf4ed3af770df01c9125dfc2d5de615772c [formerly 29c62bf4ed3af770df01c9125dfc2d5de615772c [formerly 5a4352024ae71613186730e3a3ab670f43904ad6]]
Former-commit-id: 16336c8fc22f39960ce8dd76e0c49e7cc390dacc
Former-commit-id: ee6b6e818e8568029b1ead037bca8564cf042480