lib-cvs.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d76ddb114a01881218f21115ad66571efc5a9764 [formerly 9dab19bca9ce3d3be4e704e5d230d4b1034a2cef] [formerly 9dab19bca9ce3d3be4e704e5d230d4b1034a2cef [formerly 03db917867bfd27c7a1f8e5b713d7bd7d3f6370e]]
Former-commit-id: c0a4cb57b49e1c8c834177676d8e4e6bdb22d1bd
Former-commit-id: 7182affbc754e5dceb2d091aaeeecb7865c00a47