lib-gpg.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 20de04e2f4c9141b4d50f61cb3f9ff10429a9c3d [formerly 357e35a7f371c4a176eff15b5ee7d9e8599fbd7c] [formerly 357e35a7f371c4a176eff15b5ee7d9e8599fbd7c [formerly be194d53c0ce38ab3e040c6be249550fab4e8e1c]]
Former-commit-id: b21f8509c0ece2b7f57200f09bb7981ee5c847a5
Former-commit-id: 18780a90351df4080b72d688492f00d85c732457