p5302-pack-index.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a32c99ff0c6f72465db88393c2cadcf49ae4dc90 [formerly e95c19685760ea66024ecdf7326c727b2e7d4f86] [formerly e95c19685760ea66024ecdf7326c727b2e7d4f86 [formerly 4717659144984c7fbfd2c01a2c22f342756e2abd]]
Former-commit-id: 36685c5236481b1d0350695f3af3691a04b56a5d
Former-commit-id: c31cd3a765e402e299b2a9dca03bf34d73e648e1