mergetool: document the default for --[no-]prompt
The original motivation of using the prompt was to confirm to run a
tool on this particular (as opposed to another) path, but the user
can also take the prompt as to confirm to run this (as opposed to
some other) tool.  The latter of which of course is irritating for
those who told which exact tool to use, which is the reason why we
are flipping the default.

During the review discussion of the patch, many people (including
the maintainer) missed that a user can find the prompt useful way to
skip running the tool on particular paths.  Clarify it by adding a
brief half-sentence to the description.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c48fda8b4ea5ad2fe1f5dc2986fd68d24bab281e [formerly c48fda8b4ea5ad2fe1f5dc2986fd68d24bab281e [formerly c15bb0cad73e4dad52a6d4869a254baf3a689b6a]]
Former-commit-id: 60837fa2de33d715f6575de2eb750188136b96e8
Former-commit-id: 67f5c15f88f4bd6cbd3a5bb19bf75e970e0da3f8