blame: large-scale performance rewrite
The previous implementation used a single sorted linear list of blame
entries for organizing all partial or completed work.  Every subtask had
to scan the whole list, with most entries not being relevant to the
task.  The resulting run-time was quadratic to the number of separate
chunks.

This change gives every subtask its own data to work with.  Subtasks are
organized into "struct origin" chains hanging off particular commits.
Commits are organized into a priority queue, processing them in commit
date order in order to keep most of the work affecting a particular blob
collated even in the presence of an extensive merge history.

For large files with a diversified history, a speedup by a factor of 3
or more is not unusual.

Signed-off-by: David Kastrup <dak@gnu.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3c9576b965716ed288c59bb17f69a06921f7c995 [formerly 3c9576b965716ed288c59bb17f69a06921f7c995 [formerly 7e6ac6e4391caa0fc379cb699013d503380e4214]]
Former-commit-id: 18b63440357ba6aca60f8d451b2d5b3b08615d61
Former-commit-id: 95bb5131e38f4ed02c10e6687d08404e66d8862c