replace: refactor command-mode determination
The git-replace command has three modes: listing, deleting,
and replacing. The first two are selected explicitly. If
none is selected, we fallback to listing when there are no
arguments, and replacing otherwise.

Let's figure out up front which operation we are going to
do, before getting into the application logic. That lets us
simplify our option checks (e.g., we currently have to check
whether a useless "--force" is given both along with an
explicit list, as well as with an implicit one).

This saves some lines, makes the logic easier to follow, and
will facilitate further cleanups.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Christian Couder <chriscool@tuxfamily.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1a5a08ea5971cb8bc54de2d03565246e4a5a7d34 [formerly 1a5a08ea5971cb8bc54de2d03565246e4a5a7d34 [formerly 3f495f67bc4ec744ac60f6e7bec0924022670998]]
Former-commit-id: 5da168428bb281d45d36752382ae32bb5b874d17
Former-commit-id: b0fcb83cbe8a056919709438244f6226c55d36c9