replace: use OPT_CMDMODE to handle modes
By using OPT_CMDMODE, the mutual exclusion between modes is
taken care of for us. It also makes it easy for us to
maintain a single variable with the mode, which makes its
intent more clear. We can use a single switch() to make sure
we have covered all of the modes.

This ends up breaking even in code size, but the win will be
much bigger when we start adding more modes.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Christian Couder <chriscool@tuxfamily.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ec9cfc49bcff346dd2f53dd4ae1b9e247a633ff5 [formerly ec9cfc49bcff346dd2f53dd4ae1b9e247a633ff5 [formerly 70c7bd6dafc83eee7bf76d33d12027224d80f20d]]
Former-commit-id: b29403e6c482244768e33d33a4c2d880a0c630ff
Former-commit-id: ae6cebd336705d0a79a548747a7bcf9d50123bf1