replace: add --edit option
This allows you to run:

    git replace --edit SHA1

to get dumped in an editor with the contents of the object
for SHA1. The result is then read back in and used as a
"replace" object for SHA1. The writing/reading is
type-aware, so you get to edit "ls-tree" output rather than
the binary tree format.

Missing documentation and tests.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Christian Couder <chriscool@tuxfamily.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2d3ff6fde91b6ad7cceb7bae01ebfddcc741f647 [formerly 2d3ff6fde91b6ad7cceb7bae01ebfddcc741f647 [formerly b892bb45eacb484be281a992bef66ea723210717]]
Former-commit-id: eb55605738fe2607a66b228c50e818d508b35b67
Former-commit-id: d3229a65e49f84e2069de029a8feee742720e4b1