blame: correctly handle files regardless of autocrlf
If a file contained CRLF line endings in a repository with
core.autocrlf=input, then blame always marked lines as "Not
Committed Yet", even if they were unmodified.  Don't attempt to
convert the line endings when creating the fake commit so that blame
works correctly regardless of the autocrlf setting.

Reported-by: Ephrim Khong <dr.khong@gmail.com>
Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2bd786bcc8c66559f79ff9412bbdb4924a1f9804 [formerly 2bd786bcc8c66559f79ff9412bbdb4924a1f9804 [formerly 4d4813a52f3722854a54bab046f4abfec13ef6ae]]
Former-commit-id: 854d13ffa0096512179da0213b59a86b3003edbe
Former-commit-id: 223211b63ae0172aa9adac19f91c256813d01ef4