t0010-racy-git.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: bdbeedf18a71144db0724eff52df2696ae037b27 [formerly bdbeedf18a71144db0724eff52df2696ae037b27 [formerly 8fc5593c53218d19c7a0bac34debeaf6328e4f62]]
Former-commit-id: b5620047894862dd9eea218d1e488ba1e9881e24
Former-commit-id: 2fd2793920bd765eed594dafebf593f68148ec4e