t0020-crlf.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1ea55cf3df04bbd973a974c02190d665f2442ced [formerly 01c753338757d0e4af616d42eb5679c6e8ef3037] [formerly 01c753338757d0e4af616d42eb5679c6e8ef3037 [formerly def226bdbb53d32f6c3f6bb66cfe1810089033df]]
Former-commit-id: 358e427cdfecf664cee6ab416599e6ac2ad8e769
Former-commit-id: fd3153478b2dced1e1eb0f9c9153d22270119f9e