t0025-crlf-auto.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 90760dfe1dbd137eeb61905e0da667761531d077 [formerly 90760dfe1dbd137eeb61905e0da667761531d077 [formerly 0bf6414996d5acbefaaf808ec417c45661981bd0]]
Former-commit-id: 188aadbe4e5c60885b4911e7de04d3c6ae964b55
Former-commit-id: a098118940c69d931930428915078bbd876f1eba