t0026-eol-config.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d8b01de8ab552baa9303c997e7e60d3d62dd7ba5 [formerly d8b01de8ab552baa9303c997e7e60d3d62dd7ba5 [formerly 8deeab4a2413868b44fda4a0d0c56d2a7c45c6be]]
Former-commit-id: a8204f1cecb53b964d274855e138e85b8f278c99
Former-commit-id: 1eb4a17122f8771a3337029009835fb6f59d1d74