t0030-stripspace.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 74a0d96384619360b5f425ff729fdcf9d3e027c8 [formerly 4bf7738da1e08c28cfefeb26a6ca3e2b632e1c07] [formerly 4bf7738da1e08c28cfefeb26a6ca3e2b632e1c07 [formerly 4d713567f9e79f738e7b309ad0222e1517871f9c]]
Former-commit-id: 6ff5368a42015265223ef533de1e8d16e0bea67d
Former-commit-id: e5e4fae99c2b602dcc532230f6308d28620da8a6