t0300-credentials.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 353585b1bfa3f06007ba79b67fdae06d64defd5c [formerly 353585b1bfa3f06007ba79b67fdae06d64defd5c [formerly dd64267fe2151b1ea3f19b5745d0fa8321995381]]
Former-commit-id: e6aab1f2ecd66ceef30a7879f63fdcc98f7e4f78
Former-commit-id: 039514f2c9097083097f2b4c2656b036f12db4d8