t1001-read-tree-m-2way.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 682e9433d646f4fd1b53f3b48a3c1d0a9d4488f5 [formerly 878890d540cd2db9fef6a624eee20b10cbcdac76] [formerly 878890d540cd2db9fef6a624eee20b10cbcdac76 [formerly 142efa3e43e9c61cb5b97dbe69d36ca090d683d5]]
Former-commit-id: bd475aa334d07051bc0a378e063ed991cb53fffd
Former-commit-id: fc7ed1f4ae10bb5a9521834daeab09d4f6b5a534