t1002-read-tree-m-u-2way.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b105c4cd21477f25a2c34b9d73973e2a417dd305 [formerly b105c4cd21477f25a2c34b9d73973e2a417dd305 [formerly 9b3bc877f0b01826beaa3ef00b5e3b2c49fb09cc]]
Former-commit-id: 0dffb4c050fc01bd3ab1c78768dcf1cc41ec9294
Former-commit-id: 71e087bf003b1195ceaba72fa0e97174a947d70c