t1003-read-tree-prefix.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 531f33a59770c3c3a1ef5e864ae734736f17aff9 [formerly 531f33a59770c3c3a1ef5e864ae734736f17aff9 [formerly 7f311eb54ba45e1bcdda0900da891e4652614b30]]
Former-commit-id: 843152199c1bbb19bd2d4eeb888dd7ca42f6bd78
Former-commit-id: 27990478c273351ad10404fc64454f035982684d