t1004-read-tree-m-u-wf.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2d0e1625cc2b1f785e6d9724c20bd26c92ac2630 [formerly bdbfe188ac7d1e2065da29ef50f64450826c5807] [formerly bdbfe188ac7d1e2065da29ef50f64450826c5807 [formerly 77317c0c5c24e799a397b7d32080a9490fc23b11]]
Former-commit-id: 24509ee60e00703999c9fe79305a40583acfa549
Former-commit-id: 8f588fe3a647abc59aabb50e7d1770a28cda1e7a