t1020-subdirectory.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6f57e972d4cf9e6ca4bf6817caa661d1d19bdc91 [formerly 6f57e972d4cf9e6ca4bf6817caa661d1d19bdc91 [formerly c9e454ccef2f572ce532e29b9d1317c02764545e]]
Former-commit-id: 99a5499b0ba87d3c5875cda68aeb2deabdde238b
Former-commit-id: 81d85853756ad7bdf4043ad774edeb64a4dff8c6