t1050-large.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: dff2faa12085f8db2a920813e2c86ad5176f0504 [formerly dff2faa12085f8db2a920813e2c86ad5176f0504 [formerly 714c71b2b1531f7e1099d17ab8ae03fa21f4b88c]]
Former-commit-id: 871ff00c53f36a55f237342c68c4619a10305e58
Former-commit-id: 5ad1b8f2c7314253167df84ed6fc9d716a6ec067