Makefile: default to -lcurl when no CURL_CONFIG or CURLDIR
The original implementation of CURL_CONFIG support did not match the
original behavior of using -lcurl when CURLDIR was not set. This broke
implementations that were lacking curl-config but did have libcurl
installed along system libraries, such as MSysGit. In other words, the
assumption that curl-config is always installed was incorrect.

Instead, if CURL_CONFIG is empty or returns an empty result (e.g. due
to curl-config being missing), use the old behavior of falling back to
-lcurl.

Signed-off-by: Dave Borowitz <dborowitz@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c240462a5b8405ac777383ffb2b5c2811ce4d965 [formerly e5c26ec5ecd42f690c09001795f455e0f8434615] [formerly e5c26ec5ecd42f690c09001795f455e0f8434615 [formerly f3f11fa6a5cee5a90b3ca6a66a25a2beff6996a7]]
Former-commit-id: 0fdae74ef26563532bdd6404512403aa4b816726
Former-commit-id: 707c7595153e8e669c08d77c260764f3094655e4