rerere: fix for merge.conflictstyle
If we use a different conflict style `git rerere forget` is not able
to find the matching conflict SHA-1 because the diff generated is
actually different from what `git merge` generated, due to the
XDL_MERGE_* option differences among the codepaths.

The fix is to call git_xmerge_config() so that git_xmerge_style is set
properly and the diffs match.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Reviewed-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: aef55ac0dd4e75d7fdd179f6c1724d101642473c [formerly 3e4825d37bfa6bafd310397d33c77c2af4e4bcc0] [formerly 3e4825d37bfa6bafd310397d33c77c2af4e4bcc0 [formerly de3d8bb7736cafae0ff10e047e2afa994b1dd901]]
Former-commit-id: 2298f16d82502c0e87573ab9685db71280cada3d
Former-commit-id: 44bb7fc60beea0eb9ef1ce713f4e210af5cc00fc