refs.c: make prune_ref use a transaction to delete the ref
Change prune_ref to delete the ref using a ref transaction. To do this we also
need to add a new flag REF_ISPRUNING that will tell the transaction that we
do not want to delete this ref from the packed refs. This flag is private to
refs.c and not exposed to external callers.

Signed-off-by: Ronnie Sahlberg <sahlberg@google.com>
Reviewed-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 999154beacf2cccb50f8fab5712863b3a708b1f1 [formerly 2f8a819787f57a1d9daf7f2e73247d0fb67963a9] [formerly 2f8a819787f57a1d9daf7f2e73247d0fb67963a9 [formerly 029cdb4ab21c49a916efd68eaf2d2431c7fab7c7]]
Former-commit-id: 7bd6ddd3b6182e002482620cf183affa9e984e02
Former-commit-id: ae1827b1288b19cf26686e17382c4f96c2a8d6b1