refs.c: make delete_ref use a transaction
Change delete_ref to use a ref transaction for the deletion. At the same time
since we no longer have any callers of repack_without_ref we can now delete
this function.

Change delete_ref to return 0 on success and 1 on failure instead of the
previous 0 on success either 1 or -1 on failure.

Signed-off-by: Ronnie Sahlberg <sahlberg@google.com>
Reviewed-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0a4248ac25aef781b0f2669bd04d4a1503487506 [formerly 9f415409beda6d73872f78808acd2d286c9530a6] [formerly 9f415409beda6d73872f78808acd2d286c9530a6 [formerly 7521cc4611a783f4a8174bd0fcec5f4a47357ac1]]
Former-commit-id: 3990d3d90496de36a2e19b36f738ad23f8dbd436
Former-commit-id: bbb0b6e1bc1fe87b50e99a486e07acd0febf6e29