t3905-stash-include-untracked.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: dceaba6c6bf797fa2fa44704699bf181a0308c67 [formerly dceaba6c6bf797fa2fa44704699bf181a0308c67 [formerly cba1262100df44bdb2cd5a00741562afb3eeb9fd]]
Former-commit-id: 645d1f28aa6529b0feecd5f0aceee9bcf5ecd648
Former-commit-id: ac7c1cd7ece30ae6a840f01c4c91d5e720735d74