t3910-mac-os-precompose.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f076d322f0e8f8c58771375ce5806020aab9d79d [formerly 52f8efdbdde9b2770e19f05716075cc4d3c10406] [formerly 52f8efdbdde9b2770e19f05716075cc4d3c10406 [formerly a4cf6b4b91835c1d8d13ebb5cf3ce9241e181f1b]]
Former-commit-id: 107e9c6059b4bcaba8a13771b3b0483627e3cdf4
Former-commit-id: 63c45262fda9d625e416149f66db28fd49a1b0f2