t4006-diff-mode.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 92b08f175314a4455cdb7d6ec4bc0787c40e835e [formerly c90a72c831fd143e8dccaffe7536352ddaf66af8] [formerly c90a72c831fd143e8dccaffe7536352ddaf66af8 [formerly e1d6b55d5db4c1ad305b5ac6e4f1bb4bea3c1bee]]
Former-commit-id: 639eaeebd30f4649ec3b26c70b3c19c26f58ab8a
Former-commit-id: f9720e30686dd947ad0e65b177e556b1e09e2d7e