t4010-diff-pathspec.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a6565c44b0458ca8c71dc83e33668be752dd6618 [formerly a6565c44b0458ca8c71dc83e33668be752dd6618 [formerly 38b2e5d12cbe681bff8f31cf10e3de8bd6f6db52]]
Former-commit-id: c711657b7dae79dbfdc7c5fd3d3128957f2fbb97
Former-commit-id: 154853f3f834e06568466f642290941cb877f169