t4012-diff-binary.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d0f9bb7779acc1cdcce3acbb3567df257402c963 [formerly d0f9bb7779acc1cdcce3acbb3567df257402c963 [formerly e6ce6f4c7a8bcb2a77e664d0dfc42ee9c2d164d0]]
Former-commit-id: db06c0ea6e2202fcc283b714c77db66f62b81aaa
Former-commit-id: df39a41c8bc3c923537ae9c282ff89de2aaf3e53