t4013-diff-various.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d9e6e84c27b32fe5eb0267685ec204029951499f [formerly d9e6e84c27b32fe5eb0267685ec204029951499f [formerly 4ff03347ec945542e288a6a020bdd8153c915592]]
Former-commit-id: b9b89eb752f8663d8264d5bd24b232ee8b0dc8bc
Former-commit-id: a2bdd516b2b9256eb248f1fd55df5b4130946cdc