t4014-format-patch.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b74feea144df3eaaaed2ef1cb40051e58a881954 [formerly b74feea144df3eaaaed2ef1cb40051e58a881954 [formerly 54835fc57ea16ed19d1689b37073806537f14cf7]]
Former-commit-id: 98fd562b2e03fd0fcaa442e896f4033f28511fe0
Former-commit-id: f71fb5947fcd4be68bf3b1fe9785718e21135270