t4036-format-patch-signer-mime.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: dd3e402fc542f51459449d01008d9ebc1fcd977d [formerly 25290c70eb2ade3509d7657b264f321218f5156c] [formerly 25290c70eb2ade3509d7657b264f321218f5156c [formerly 20cb28baf9d03c779f590baa475610788bc6eabf]]
Former-commit-id: e31192183bb8ee2b88469d95cb9001b556428111
Former-commit-id: 2de3418d038091d991346c7b8fdabaeaf2bb72ee