t4038-diff-combined.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7ebf88d4e4841146289bae9e6cd219a03bcdcce2 [formerly 4f4e59579380d6141316eeb4df125fa4058ffb60] [formerly 4f4e59579380d6141316eeb4df125fa4058ffb60 [formerly 7c0c51baa4855c76a71820fb279c2ca1580d374b]]
Former-commit-id: 83a0c5e8e1ee7896e3bea1bdbeea78ed5da4d260
Former-commit-id: f1a064452134c38285439967c7498478b6a078fb