t4057-diff-combined-paths.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d41cfd41d8383ce73efb39e08d1690ca16c94f37 [formerly d41cfd41d8383ce73efb39e08d1690ca16c94f37 [formerly 274447aa6bbe84c4e65607f48df18380766d24fe]]
Former-commit-id: 8cc3c31e638e36846c07718456ad2120b8ddd045
Former-commit-id: b09e68a161fd29f0e1309c462381eea24327c5da