t4116-apply-reverse.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b48182866ba2c7ad59543ad5385140309c65ca7b [formerly b48182866ba2c7ad59543ad5385140309c65ca7b [formerly 991a9c3af984eb28b31f8642c97fd9ddb29cadc4]]
Former-commit-id: 8e6ef762441669f35320a213cbbef981ad12e934
Former-commit-id: 89e7af7a212c4db044ab4b4af25a5eeb68190751