t4119-apply-config.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4b20c2916156a37324f9602add52388fe1fee995 [formerly 4b20c2916156a37324f9602add52388fe1fee995 [formerly 6003eb13c69516dfedc9466aeb8c3a32a054c286]]
Former-commit-id: f4c9d49dd3ed48dec954fe5d4ec159be27e37ed4
Former-commit-id: a1af317d50afa36a5cbf31629623e2405adad608