t4204-patch-id.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 03ab02ea77fb128a250d9f5cd7f50b5428f0e97d [formerly 03ab02ea77fb128a250d9f5cd7f50b5428f0e97d [formerly ce21ccfae0729fd73b54f3ae4935fc7769343561]]
Former-commit-id: 50db2c6ebb15d1521db03ccdfc05a8dd26271a0a
Former-commit-id: c2140f511655e6b37d85672788d0e6d8d1ec1033