t5000-tar-tree.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

for _f in $(find . -name "*.sh")
do
   sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>
Reviewed-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c62d5e00670fa57f1d6ae8f0587df6d9f1940367 [formerly b0d0676817b7ef3b6b999a655f4ae4ac3a872758] [formerly b0d0676817b7ef3b6b999a655f4ae4ac3a872758 [formerly f5efd5196cf2d7931785f663b219b40fd204e1b1]]
Former-commit-id: 51005a766ed84c7f7437c58e8d7113c60171a6cb
Former-commit-id: f4f14d0ac2935b07044ea22d4444881879584bc0