merge-recursive.c: fix case-changing merge bug
On a case-insensitive filesystem, when merging, a file would be
wrongly deleted from the working tree if an incoming commit had
renamed it changing only its case.  When merging a rename, the file
with the old name would be deleted -- but since the filesystem
considers the old name to be the same as the new name, the new
file would in fact be deleted.

We avoid this by not deleting files that have a case-clone in the
index at stage 0.

Signed-off-by: David Turner <dturner@twitter.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: bd1375e632bb71661e5b840545ee57a97c160b31 [formerly bd1375e632bb71661e5b840545ee57a97c160b31 [formerly ae352c7f37ef2098e03ee86bc7fd75b210b17683]]
Former-commit-id: c7125b9792fa6c33cc0a7b82ee77dab431680d39
Former-commit-id: f59af047e2611c4f6b788b59426c20be20091854