commit: print "Date" line when the user has set date
When we make a commit and the author is not the same as the
committer (e.g., because you used "-c $commit" or
"--author=$somebody"), we print the author's name and email
in both the commit-message template and as part of the
commit summary. This is a safety check to give the user a
chance to confirm that we are doing what they expect.

This patch brings the same safety for the "date" field,
which may be set by "-c" or by using "--date".  Note that we
explicitly do not set it for $GIT_AUTHOR_DATE, as it is
probably not of interest when "git commit" is being fed its
parameters by a script.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d89a3c9af3c49708bf4a5accad741c766bdc0c52 [formerly d89a3c9af3c49708bf4a5accad741c766bdc0c52 [formerly b7242b8c9e4b3c57a07c2a76d0337389605aadcc]]
Former-commit-id: 80f16455d4bccbcb37e933a19d4fba97f2234c8a
Former-commit-id: 9106a92ae77f9695e47f4d31a24c329a808127f2