commit: accept more date formats for "--date"
Right now we pass off the string found by "--date" straight
to the fmt_ident function, which will use our strict
parse_date to normalize it. However, this means obvious
things like "--date=now" or "--date=2.days.ago" will not
work.

Instead, let's fallback to the approxidate function to
handle this for us. Note that we must try parse_date
ourselves first, even though approxidate will try strict
parsing itself. The reason is that approxidate throws away
any timezone information it sees from the strict parsing,
and we want to preserve it. So asking for:

  git commit --date="@1234567890 -0700"

continues to set the date in -0700, regardless of what the
local timezone is.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e82b037f999c85ab46aa175c21b381b77f3e7740 [formerly e82b037f999c85ab46aa175c21b381b77f3e7740 [formerly 14ac2864dcd566a025e449ab9db6b5dc57744a32]]
Former-commit-id: 5c2df3727608f951199d98f719f61e016158fc12
Former-commit-id: fa05f069da10b99e49f945640561b55632c08f9f