compat/bswap.h: restore preference __BIG_ENDIAN over BIG_ENDIAN
The previous commit swaps the order we check the macros defined by
the compiler and the system headers from the original.  Since the
order of check should not matter (i.e. it is insane to define both
__BIG_ENDIAN and friends and BIG_ENDIAN and friends and in a
conflicting way), it is the most conservative thing to do not to
change it.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b03fc80939230981f8931af888b680443d612eae [formerly b99fd1ddc36cb07dbd298d5589f84dee1e45e2cc] [formerly b99fd1ddc36cb07dbd298d5589f84dee1e45e2cc [formerly 839fa9c500c7771d15831e2bcf77697cd70620f9]]
Former-commit-id: ce87e2f763751fa82b4ebb7729db1aa7ce8c4ca1
Former-commit-id: 757709ab73e5b6592c8f2d510b61a305b68571cf