inline constant return from error() function
Commit e208f9c introduced a macro to turn error() calls
into:

  (error(), -1)

to make the constant return value more visible to the
calling code (and thus let the compiler make better
decisions about the code).

This works well for code like:

  return error(...);

but the "-1" is superfluous in code that just calls error()
without caring about the return value. In older versions of
gcc, that was fine, but gcc 4.9 complains with -Wunused-value.

We can work around this by encapsulating the constant return
value in a static inline function, as gcc specifically
avoids complaining about unused function returns unless the
function has been specifically marked with the
warn_unused_result attribute.

We also use the same trick for config_error_nonbool and
opterror, which learned the same error technique in a469a10.

Reported-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6d11ec5d4c37b81118538fcd473e6d32a7f75fe5 [formerly 6d11ec5d4c37b81118538fcd473e6d32a7f75fe5 [formerly 87fe5df3653cf20b6bf9854bea42e4016c7d4688]]
Former-commit-id: a839156f8d944f57f65c30c3ee015abeb2c1a72a
Former-commit-id: 3b7e8126a7c28955f9add09ac2d456b1302f3b1e