let clang use the constant-return error() macro
Commit e208f9c converted error() into a macro to make its
constant return value more apparent to calling code.  Commit
5ded807 prevents us using this macro with clang, since
clang's -Wunused-value is smart enough to realize that the
constant "-1" is useless in some contexts.

However, since the last commit puts the constant behind an
inline function call, this is enough to prevent the
-Wunused-value warning on both modern gcc and clang. So we
can now re-enable the macro when compiling with clang.

Tested with clang 3.3, 3.4, and 3.5.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 86243ba0b091517713c95820128d58d0665fc03a [formerly 86243ba0b091517713c95820128d58d0665fc03a [formerly ff0a80af724e81dbad6a269847523e39c2e7e479]]
Former-commit-id: 27190c2ce77d7e86d592ec09da99ca68d4b1048c
Former-commit-id: 4d0ce5f4b53bf99e1551d4fe72517b5df7064ef3