refs.c: add new functions reflog_exists and delete_reflog
Add two new functions, reflog_exists and delete_reflog, to hide the internal
reflog implementation (that they are files under .git/logs/...) from callers.
Update checkout.c to use these functions in update_refs_for_switch instead of
building pathnames and calling out to file access functions. Update reflog.c
to use these to check if the reflog exists. Now there are still many places
in reflog.c where we are still leaking the reflog storage implementation but
this at least reduces the number of such dependencies by one. Finally
change two places in refs.c itself to use the new function to check if a ref
exists or not isntead of build-path-and-stat(). Now, this is strictly not all
that important since these are in parts of refs that are implementing the
actual file storage backend but on the other hand it will not hurt either.

Signed-off-by: Ronnie Sahlberg <sahlberg@google.com>
Acked-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6ad0f40fdd014416cdf3083c481b030aa6126008 [formerly 88069feca889e5ea0722f261a45494374ae746be] [formerly 88069feca889e5ea0722f261a45494374ae746be [formerly 4da588357a4a8b73f6a8d9c24435dabee74d0a7e]]
Former-commit-id: 6af5b37a52845b8339de9f4319c7aedf29b8c847
Former-commit-id: b070891eb5347f1678ae5f648fdec72253acd029