grep: use run-command's "dir" option for --open-files-in-pager
Git generally changes directory to the repository root on
startup.  When running "grep --open-files-in-pager" from a
subdirectory, we chdir back to the original directory before
running the pager, so that we can feed the relative
pathnames to the pager.

We currently do this chdir manually, but we can ask
run_command to do it for us. This is fewer lines of code,
and as a bonus, the chdir is limited to the child process,
which avoids any unexpected surprises for code running after
the pager (there isn't any currently, but this is
future-proofing).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ce1c32ababa9854f213b0ecd8f199638f248f832 [formerly ae062c571d9607be7ff61e7857dcff80a514bbbd] [formerly ae062c571d9607be7ff61e7857dcff80a514bbbd [formerly 26ecfe3e2041fe651c9682ebb73911777290ac62]]
Former-commit-id: c25e89a656e9aed7a56c7f562650b1cb18aa7fcf
Former-commit-id: 1da8d4b7649d3dec8f6898f591a18a59c7061bb1