git-p4: format-patch to diff-tree change breaks binary patches
When applying binary patches a full index is required. format-patch
already handles this, but diff-tree needs '--full-index' argument
to always output full index. When git-p4 runs git-apply to test
the patch, git-apply rejects the patch due to abbreviated blob
object names. This is the error message git-apply emits in this
case:

    error: cannot apply binary patch to '<filename>' without full index line
    error: <filename>: patch does not apply

Signed-off-by: Tolga Ceylan <tolga.ceylan@gmail.com>
Acked-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c8f612825db88917b7073c2b4b03515cde896d21 [formerly c8f612825db88917b7073c2b4b03515cde896d21 [formerly 749b668c7db563609de3de7595504f2dca8eef7d]]
Former-commit-id: c5b1c3429cb73db5b5ffa80c7b8b25f305cdb1f6
Former-commit-id: a7c57695a0e7d660a9d4d32e149aa78ecaaa67f7