mv: allow renaming to fix case on case insensitive filesystems
"git mv hello.txt Hello.txt" on a case insensitive filesystem
always triggers "destination already exists" error, because these
two names refer to the same path from the filesystem's point of
view, and requires the user to give "--force" when correcting the
case of the path recorded in the index and in the next commit.

Detect this case and allow it without requiring "--force".

Signed-off-by: David Turner <dturner@twitter.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8b338954675899f6afc09ac15fa7c34f8f8b3ee4 [formerly 8b338954675899f6afc09ac15fa7c34f8f8b3ee4 [formerly baa37bff9a845471754d3f47957d58a6ccc30058]]
Former-commit-id: f0e5932ff5bdb1d7d4c4e9c59fc450e7a099f05a
Former-commit-id: 729914bfec511f3141cc6881a99db9446ffb3172