contrib: remove vim support instructions
The git support scripts started shipping in upstream vim in version
7.2 (2008-08-09).  Clean up contrib/ a little by removing the
instructions for people on older versions of vim.

RHEL 6 already has vim 7.2.something, so anyone on a reasonably modern
operating system should not be affected.  Users on RHEL 5 presumably
know that means sometimes missing out on niceties like syntax
highlighting, so this should be safe.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 63332c7a303f5349b8a47b0e460c071c1edfd890 [formerly 63332c7a303f5349b8a47b0e460c071c1edfd890 [formerly 7d445f518ef3b38a26166d1ee5a4074f15bc7cd1]]
Former-commit-id: 93579d36c2f48c15c0758c1ce63ce8a54b9b0407
Former-commit-id: 60fa895bc056645585bc29984ee1a14dcabcad41