contrib: remove git-diffall
The functionality of the "git diffall" script in contrib/ was
incorporated into "git difftool" when the --dir-diff option was added
in v1.7.11 (ca. June, 2012).  Once difftool learned those features,
the diffall script became obsolete.

The only difference in behavior is that when comparing to the working
tree, difftool copies any files modified by the user back to the
working tree when the diff tool exits.  "git diffall" required the
--copy-back option to do the same.  All other diffall options have the
same meaning in difftool.

Make life easier for people choosing a tool to use by removing the old
diffall script.  A pointer in the release notes should be enough to
help current users migrate.

Helped-by: Tim Henigan <tim.henigan@gmail.com>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6b92d2c18e2d8b162d78263b75c206fea44e21d2 [formerly 6b92d2c18e2d8b162d78263b75c206fea44e21d2 [formerly 502b0a1ad145a7d77733008946c833ba7d3d100d]]
Former-commit-id: 1bb14b0cce61bca53505e4e501d0c56fe3bfc4a8
Former-commit-id: 5b3bcd2ca66d58febc3eef2c625d792157c28058