t/lib-httpd: require SANITY prereq
Our test httpd setup will not generally run as root, because
Apache will want to setuid, and we do not set up the "User"
config directive. On some systems, like current Debian
unstable, Apache fails to start, and we skip the tests:

    $ sudo ./t5539-fetch-http-shallow.sh --debug
    1..0 # SKIP web server setup failed
    $ cat trash*t5539*/httpd/error.log
    [...]
    (22)Invalid argument: AH00024: Couldn't set permissions on
      the rewrite-map mutex; check User and Group directives
    AH00016: Configuration Failed

However, on other systems (reportedly Ubuntu 11.04), Apache
seems to start, and then bails during our tests with:

   getpwuid: couldn't determine user name from uid 4294967295,
     you probably need to modify the User directive
   Child 12037 returned a Fatal error...  Apache is exiting!

This may be related to the pre-fork/threading model in use
(note that the second one complains of the child dying).
However, it's not even worth investigating; in either case
we just want to skip the tests, and we already recommend
against running the test suite as root. Let's just
explicitly check this condition and skip the tests rather
than expecting Apache to do the right thing.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: edadbcb0638040bc547dc3bf0b0fd4e7f3617ace [formerly 6c3b18cadf7e8442e8869567014a17b967af389d] [formerly 6c3b18cadf7e8442e8869567014a17b967af389d [formerly a1a301114e0ba3f60474d857c93d02890d00e65b]]
Former-commit-id: cda2ccc6005f62c24ba62f4d8c0cf80ed620ab66
Former-commit-id: d2ad7334dea3aef85d6c0620c06f762e311bd1f2