contrib/subtree: allow adding an annotated tag
cmd_add_commit() is passed FETCH_HEAD by cmd_add_repository, which
is then rev-parsed into an object name.  However, if the user is
fetching a tag rather than a branch HEAD, such as by executing:

  $ git subtree add -P oldGit https://github.com/git/git.git tags/v1.8.0

the object name refers to a tag and is never peeled, and the git
commit-tree call (line 561) slaps us in the face because it doesn't
peel tags to commits.

Because peeling a committish doesn't do anything if it's already a
commit, fix by peeling the object name before assigning it to $rev
using peel_committish() from git:git-sh-setup.sh, a pre-existing
dependency of git-subtree.

Reported-by: Kevin Cagle <kcagle@micron.com>
Helped-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: James Denholm <nod.helm@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0b510eabefbbd67b3697ec9755e81cfbc367f0d5 [formerly 0b510eabefbbd67b3697ec9755e81cfbc367f0d5 [formerly 6f1871fe0f7d9bddfbe199cceeb255cf8edbaab3]]
Former-commit-id: 6ef2fc9ca16f47c0bedcbb9a3a96199c519a58b6
Former-commit-id: 87aa39d8eea7e190166baf1f881c5c2d9d611f09