run-command: store an optional argv_array
All child_process structs need to point to an argv. For
flexibility, we do not mandate the use of a dynamic
argv_array. However, because the child_process does not own
the memory, this can make memory management with a
separate argv_array difficult.

For example, if a function calls start_command but not
finish_command, the argv memory must persist. The code needs
to arrange to clean up the argv_array separately after
finish_command runs. As a result, some of our code in this
situation just leaks the memory.

To help such cases, this patch adds a built-in argv_array to
the child_process, which gets cleaned up automatically (both
in finish_command and when start_command fails).  Callers
may use it if they choose, but can continue to use the raw
argv if they wish.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d6011ab08b0b0f1e5f6372429db4e7b618896ee3 [formerly d6011ab08b0b0f1e5f6372429db4e7b618896ee3 [formerly c460c0ecdca46be85f6d9c845f9df7ce0e45c3c2]]
Former-commit-id: 91e63ecd2b397c40569f64609f70ff095d27d782
Former-commit-id: 42365c57e17965ea4b25c243c8dbcd486def6fc2