run_column_filter: use argv_array
We currently set up the argv array by hand in a fixed-size
stack-local array. Using an argv array is more readable, as
it handles buffer allocation us (not to mention makes it
obvious we do not overflow the array).

However, there's a more subtle benefit, too. We leave the
function having run start_command (with the child_process
in a static global), and then later run finish_command from
another function. That means when we run finish_command,
neither column_process.argv nor the memory it points to is
valid any longer.

Most of the time finish_command does not bother looking at
argv, but it may if it encounters an error (e.g., waitpid
failure or signal death). This is unusual, which is why
nobody has noticed. But by using run-command's built-in
argv_array, the memory ownership is handled for us
automatically.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2b31ecbd73d1c82436d09751c5bbf3fbfc21ae89 [formerly 2b31ecbd73d1c82436d09751c5bbf3fbfc21ae89 [formerly 5eb7f7ead8537420fde3eb344640096fc330bc16]]
Former-commit-id: 36511e7607e8462908035b173c18a9f9e744ea9b
Former-commit-id: 012fdba10e0448e22acc13f8a741b496b6b3e975