get_helper: use run-command's internal argv_array
The get_helper functions dynamically allocates an
argv_array, feeds it to start_command, and then returns. We
then have to later clean up the memory manually after
calling finish_command. We can make this simpler by just
using run-command's internal argv_array, which handles
cleanup for us.

This also prevents a memory leak in the case that
transport_take_over is used, in which case we free the child
in finish_connect, which does not manually free the array.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7f30a688df80aaeec4e5e8b8eff272fc9df62546 [formerly c7a14b4a218379b79f156c80844d10ef6ed572af] [formerly c7a14b4a218379b79f156c80844d10ef6ed572af [formerly e0ab2ac6c553cbba5d0275cfd35beb3351cae034]]
Former-commit-id: 3b180093d6ae4fbb17d8716c2491eb2c3e9a31ae
Former-commit-id: 25df4d8b9db707025c134e9572e5b446b0a338e5