request-pull: resurrect for-linus -> tags/for-linus DWIM
Older versions of Git before v1.7.10 did not DWIM

    $ git pull $URL for-linus

to the tag "tags/for-linus" and the users were required to say

    $ git pull $URL tags/for-linus

instead.  Because newer versions of Git works either way,
request-pull used to show tags/for-linus when asked

    $ git request-pull origin/master $URL for-linus

The recent updates broke this and in the output we see "for-linus"
without the "tags/" prefix.

As v1.7.10 is more than 2 years old, this should matter very little
in practice, but resurrecting it is very simple.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a4605331f057efe1acd86be3fe196622e38a83b8 [formerly a4605331f057efe1acd86be3fe196622e38a83b8 [formerly d952cbb1908b54c6a04f37c8b0924a49cdb5997d]]
Former-commit-id: f92f44e419cff054b3d6e1b241094db69f19dad5
Former-commit-id: fcd7080140c0b0cdd7bdf6ff45a60c1b941ef00b