config: be strict on core.commentChar
We don't support comment _strings_ (at least not yet). And multi-byte
character encoding could also be misinterpreted.

The test with two commas is updated because it violates this. It's
added with the patch that introduces core.commentChar in eff80a9
(Allow custom "comment char" - 2013-01-16). It's not clear to me _why_
that behavior is wanted.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 830dec3edaf0c3cc40e2c57532bda2bea9a0ee45 [formerly 05f25deee0fecf8f7e758419a0c3afe015ebadc7] [formerly 05f25deee0fecf8f7e758419a0c3afe015ebadc7 [formerly 50b54fd72a8776d628365c7760bc922902d29d11]]
Former-commit-id: 8554a293364f7619c5aafe91f7caceece5402bef
Former-commit-id: 96f8aa1d0d309d667b9a46744af0cc606c03a0b3