commit: allow core.commentChar=auto for character auto selection
When core.commentChar is "auto", the comment char starts with '#' as
in default but if it's already in the prepared message, find another
char in a small subset. This should stop surprises because git strips
some lines unexpectedly.

Note that git is not smart enough to recognize '#' as the comment char
in custom templates and convert it if the final comment char is
different. It thinks '#' lines in custom templates as part of the
commit message. So don't use this with custom templates.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e5d4d9e2fda23da7781401362e0c2aa450533c00 [formerly e5d4d9e2fda23da7781401362e0c2aa450533c00 [formerly 84c9dc2c5a2d34351a06554af32501d4f99990e9]]
Former-commit-id: 0b0ef8bc900ad2235886bb54d9d3bc76a9e10540
Former-commit-id: 43698c6763951a50f883d8a997e939ab10803752