git-gui: tolerate major version changes when comparing the git version
Since git 2.0.0 starting git gui in a submodule using a gitfile fails with
the following error:

   No working directory ../../../<path>

   couldn't change working directory
   to "../../../<path>": no such file or
   directory

This is because "git rev-parse --show-toplevel" is only run when git gui
sees a git version of at least 1.7.0 (which is the version in which the
 --show-toplevel option was introduced). But "package vsatisfies" returns
false when the major version changes, which is not what we want here.

Fix that for both places where the git version is checked using vsatisfies
by appending a '-' to the version number. This tells vsatisfies that a
change of the major version is not considered to be a problem, as long as
the new major version is larger. This is done for both the place that
caused the reported bug and another spot where the git version is tested
for another feature.

Reported-by: Chris Packham <judge.packham@gmail.com>
Reported-by: Yann Dirson <ydirson@free.fr>
Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>

Former-commit-id: 773b42879a761d137550ef1b49d979458f4b32e3 [formerly 773b42879a761d137550ef1b49d979458f4b32e3 [formerly 3decb8e0ac5d951f0a0cf6e604790b471aef22e1]]
Former-commit-id: 1aedcafd0ca557000eceb2519326e967ebb8ec8f
Former-commit-id: e2517f63117053af87eaaa760d67d2f2c659f1d5