refs.c: make ref_transaction_begin take an err argument
Add an err argument to _begin so that on non-fatal failures in future ref
backends we can report a nice error back to the caller.
While _begin can currently never fail for other reasons than OOM, in which
case we die() anyway, we may add other types of backends in the future.
For example, a hypothetical MySQL backend could fail in _begin with
"Can not connect to MySQL server. No route to host".

Signed-off-by: Ronnie Sahlberg <sahlberg@google.com>
Reviewed-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6d7190e26ec8251c16ec7299fff0fc753596d47a [formerly 6d7190e26ec8251c16ec7299fff0fc753596d47a [formerly 93a644ea9d3702cc1cc62c0d413f81f8e46fabe7]]
Former-commit-id: 371300f32f687ef3605c6629cdb9fdcd7b8cba70
Former-commit-id: 434f86c89510e88283720dc26b71dda069373cc5