rebase -i: handle "Nothing to do" case with autostash
When a user invokes

  $ git rebase -i @~3

with dirty files and rebase.autostash turned on, and exits the $EDITOR
with an empty buffer, the autostash fails to apply. Although the primary
focus of rr/rebase-autostash was to get the git-rebase--backend.sh
scripts to return control to git-rebase.sh, it missed this case in
git-rebase--interactive.sh. Since this case is unlike the other cases
which return control for housekeeping, assign it a special return status
and handle that return value explicitly in git-rebase.sh.

Reported-by: Karen Etheridge <ether@cpan.org>
Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f5f56ff35fb01b4c0200a1266e51dbce4cfd5124 [formerly f5f56ff35fb01b4c0200a1266e51dbce4cfd5124 [formerly e4244eb3957d478ca05b42d8b56aadb9ab2bc7ae]]
Former-commit-id: 614c58211be2ec2f706dcd8287c73732733078c6
Former-commit-id: 18f07a166137d67991a48333469aef714521c7ee