move "--follow needs one pathspec" rule to diff_setup_done
Because of the way "--follow" is implemented, we must have
exactly one pathspec. "git log" enforces this restriction,
but other users of the revision traversal code do not. For
example, "git format-patch --follow" will segfault during
try_to_follow_renames, as we have no pathspecs at all.

We can push this check down into diff_setup_done, which is
probably a better place anyway. It is the diff code that
introduces this restriction, so other parts of the code
should not need to care themselves.

Reported-by: "Michael S. Tsirkin" <mst@redhat.com>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2034c11918c5032f97014eb65f2552a58b23d1a6 [formerly 2034c11918c5032f97014eb65f2552a58b23d1a6 [formerly dd63f169d9381a77db469325688a2bef1a79a9e1]]
Former-commit-id: 89dbe3e983d30452b03074f2ba31c3caea8b6dd1
Former-commit-id: ad71a1ad62fac9b7c067cca6eae71ab698e607c7