t4205, t6006: add tests that fail with i18n.logOutputEncoding set
Pretty format string %<(N,[ml]trunc)>%s truncates subject to a given
length with an appropriate padding. This works for non-ASCII texts when
i18n.logOutputEncoding is UTF-8 only (independently of a printed commit
message encoding) but does not work when i18n.logOutputEncoding is NOT
UTF-8.

There were no breakages as far as were no tests for the case
when both a commit message and logOutputEncoding are not UTF-8.

Add failing tests for that which will be fixed in the next patch.

Signed-off-by: Alexey Shumkin <Alex.Crezoff@gmail.com>
Helped-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Helped-by: Ramsay Jones <ramsay@ramsay1.demon.co.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3299a1502b3f7a8b1a4b59e76bcffb4563972ac1 [formerly a349f3d11278939cdc8f41a363981d5690e29ba1] [formerly a349f3d11278939cdc8f41a363981d5690e29ba1 [formerly d928d81051fc8475c7845526b255cb8e80b39f12]]
Former-commit-id: 0a036835dab72e9b7c196e34e89aeeaf4e0497b0
Former-commit-id: 3606a31c78580b122621547e0601da413544d04a