format-patch: make newline after signature conditional
When we print an email signature, we print the divider
"-- \n", then the signature string, then two newlines.

Usually the signature is a one-liner (and the default is just the
git version), so the extra newline makes sense.  But one could
easily specify a multi-line signature, like this:

  git format-patch --signature='this is my long signature

  it has multiple lines
  ' ...

and it may end with its own newline, in which case we do not have
to add yet another one.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Jeremiah Mahler <jmmahler@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3456c40ba02ee324918493ce345e33eb2c151deb [formerly 347bf21ecf50e47d8cf4ebd9ca09f7c3b118c4cb] [formerly 347bf21ecf50e47d8cf4ebd9ca09f7c3b118c4cb [formerly c6076e2b4aaebd4a49b01255fe8c67405704aa72]]
Former-commit-id: 38b87ed7b4786a0102c160ecb9f3f86d62f33d47
Former-commit-id: 6d192af7da98c579a0427202e227cc5b3ad0691f