http: extract type/subtype portion of content-type
When we get a content-type from curl, we get the whole
header line, including any parameters, and without any
normalization (like downcasing or whitespace) applied.
If we later try to match it with strcmp() or even
strcasecmp(), we may get false negatives.

This could cause two visible behaviors:

  1. We might fail to recognize a smart-http server by its
     content-type.

  2. We might fail to relay text/plain error messages to
     users (especially if they contain a charset parameter).

This patch teaches the http code to extract and normalize
just the type/subtype portion of the string. This is
technically passing out less information to the callers, who
can no longer see the parameters. But none of the current
callers cares, and a future patch will add back an
easier-to-use method for accessing those parameters.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9705e49ffb30a9b819ee3694585023cf84c77873 [formerly 9705e49ffb30a9b819ee3694585023cf84c77873 [formerly bf197fd7eebcb3579dd659af35822ce88adc66c8]]
Former-commit-id: cbb397b3a645df3b2136dc4ed2b84d9c79e9da43
Former-commit-id: 4b73c05e5dbf8ada3e620fa03f2ef88e7762a213