http: optionally extract charset parameter from content-type
Since the previous commit, we now give a sanitized,
shortened version of the content-type header to any callers
who ask for it.

This patch adds back a way for them to cleanly access
specific parameters to the type. We could easily extract all
parameters and make them available via a string_list, but:

  1. That complicates the interface and memory management.

  2. In practice, no planned callers care about anything
     except the charset.

This patch therefore goes with the simplest thing, and we
can expand or change the interface later if it becomes
necessary.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: afd4f4e83d8d326293e9c4ba64b9b8faccdc62eb [formerly afd4f4e83d8d326293e9c4ba64b9b8faccdc62eb [formerly e31316263af98c4583be39b469f3152a23eba91d]]
Former-commit-id: aa5cee27e0c8ace69b84af7ddc1ed7b5c5651615
Former-commit-id: 01bd9bddf8e49225e602df1000c4f819df7fd49f