test-lint: find unportable sed, echo, test, and export usage after &&
Instead of anchoring these checks with "^\s*", just check that the
usage is preceded by a word boundary.  So now we can catch

	test $cond && export foo=bar

just like we already catch

	test $cond &&
	export foo=bar

As a side effect, this will detect usage of "sed -i", "echo -n", "test
a == b", and "export a=b" in comments.  That is not ideal but it's
potentially useful because people sometimes copy code from comments so
it can be good to also avoid nonportable patterns there.

To avoid false positives, keep the checks for 'declare' and 'which'
anchored.  Those are frequently used words in normal English-language
comments.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 77f2ef349c976e837fe465158b75ae0e889e48b2 [formerly ba43f217eaef9e7e1de6f7c36277b6c2fbc2095d] [formerly ba43f217eaef9e7e1de6f7c36277b6c2fbc2095d [formerly 561b46c5c8a9d204974e6edd4f58a5fca343d6d3]]
Former-commit-id: e26fc52389f4868f2cf3f6c6c9c9f2088e9616a9
Former-commit-id: a2276e1077d1e3d6b1f13894850f9599eac30f64