strbuf: add strbuf_tolower function
This is a convenience wrapper to call tolower on each
character of the string.

This makes config's lowercase() function obsolete, though
note that because we have a strbuf, we are careful to
operate over the whole strbuf, rather than assuming that a
NUL is the end-of-string.

We could continue to offer a pure-string lowercase, but
there would be no callers (in most pure-string cases, we
actually duplicate and lowercase the duplicate, for which we
have the xstrdup_tolower wrapper).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: dd3666fca1e14c568b190b2c2d7238b79861e282 [formerly dd3666fca1e14c568b190b2c2d7238b79861e282 [formerly ffb20ce125f0ffe59002b1258b06c26747b69b7b]]
Former-commit-id: af426bcb07d734756e04acb6355830536a0abb2a
Former-commit-id: 04ff0244c7c6ba69ea68aa7ea597459a35f54f9c