git-p4: Do not include diff in spec file when just preparing p4
The diff information render the spec file unusable as is by p4,
do not include it when run with --prepare-p4-only so that the
given file can be directly passed to p4.

With --prepare-p4-only, git-p4 already tells the user it can use
p4 submit with the generated spec file. This fails because of the
diff being present in the file. Not including the diff fixes that.

Without --prepare-p4-only, keeping the diff makes sense for a
quick review of the patch before submitting it. And does not cause
problems with p4 as we remove it programmatically.

Signed-off-by: Maxime Coste <frrrwww@gmail.com>
Acked-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4b6cb686b12f7d767680591bbedf06ed019183d2 [formerly 3abba35ee91bf1c7c002fe9cfff953ddeda85857] [formerly 3abba35ee91bf1c7c002fe9cfff953ddeda85857 [formerly b4073bb387ef303c9ac3c044f46d6a8ae6e190f0]]
Former-commit-id: 9e1b1829f5abb98e032814c4fe6b35292d6dc839
Former-commit-id: 626da9272c72847246fc94df2c5ada1ac4e445a4