fetch doc: remove notes on outdated "mixed layout"
In old days before Git 1.5, it was customery for "git fetch" to use
the same local branch namespace to keep track of the remote-tracking
branches, and it was necessary to tell users not to check them out
and commit on them.  Since everybody uses the separate remote layout
these days, there is no need to warn against the practice to check
out the right-hand side of <refspec> and build on it---the RHS is
typically not even a local branch.

Incidentally, this also kills one mention of "Pull:" line of
$GIT_DIR/remotes/* configuration, which is a lot less familiar to
new people than the more modern remote.*.fetch configuration
variable.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e500456f60c4cc1a6e65f907975a9b6bdf1fb186 [formerly e500456f60c4cc1a6e65f907975a9b6bdf1fb186 [formerly 3630654956b75b10ae19211599e08257cd4c3f8d]]
Former-commit-id: 6290edd4d8558abb558e61f047d52b6f42489aba
Former-commit-id: b0dfebee81312a657d7f73b88bcf1794a6cae24f