fetch doc: remove "short-cut" section
It is misleading to mention that <ref> that does not store is to
fetch the ref into FETCH_HEAD, because a refspec that does store is
also to fetch the LHS into FETCH_HEAD.  It is doubly misleading to
list it as part of "short-cut".  <ref> stands for a refspec that has
it on the LHS with a colon and an empty RHS, and that definition
should be given at the beginning of the entry where the format is
defined.

Tentatively remove this misleading description, which leaves the
`tag <tag>` as the only true short-hand, so move it at the beginning
of the entry.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e6ada8a4139192d3e85603a8f1c12487cc8b703c [formerly e6ada8a4139192d3e85603a8f1c12487cc8b703c [formerly 5cc3268720bd64f7dcc11843060fcebbe84f1a39]]
Former-commit-id: ab060088c729addb4e1ed253db55dcb368ee716a
Former-commit-id: 83846bae0ccdef39aa6b14cc877ae803d1b56c85