fetch: allow explicit --refmap to override configuration
Since the introduction of opportunisitic updates of remote-tracking
branches, started at around f2690487 (fetch: opportunistically
update tracking refs, 2013-05-11) with a few updates in v1.8.4 era,
the remote.*.fetch configuration always kicks in even when a refspec
to specify what to fetch is given on the command line, and there is
no way to disable or override it per-invocation.

Teach the command to pay attention to the --refmap=<lhs>:<rhs>
command-line options that can be used to override the use of
configured remote.*.fetch as the refmap.

Signed-off-by: Junio C Hamano <gitster@pobox.com>
---

Former-commit-id: ccdb74c9e95972ce43cd1703be987261be7f6538 [formerly f5ce5d6b87c229ae7041579f5348b9d1dc361930] [formerly f5ce5d6b87c229ae7041579f5348b9d1dc361930 [formerly c5558f80c3a34f411022f1bc67f8364556ca3922]]
Former-commit-id: 76a5a6feab7e334b46f756d20dbb999e78de0781
Former-commit-id: beeaa8b7469ca0c91bb7bc4194ee1f7a6280d46a