compat/bswap.h: fix endianness detection
The changes to make detection of endianness more portable had a bug
that breaks on (at least) Solaris x86.

The bug appears to be a simple copy/paste typo. It checks for
_BIG_ENDIAN and not _LITTLE_ENDIAN for both the case where we would
decide the system is big endian and little endian. Instead, the
second test should be for _LITTLE_ENDIAN and not _BIG_ENDIAN.

Two fixes were possible:

 1. Change the negation order of the conditions in the second test.
 2. Reverse the order of the conditions in the second test.

Use the second option so that the condition we expect is always a
positive check.

Signed-off-by: Ben Walton <bdwalton@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: af47a14a50aeccc1db5b1e1e6efedec69c393ba6 [formerly af47a14a50aeccc1db5b1e1e6efedec69c393ba6 [formerly 9c65ee15ee11c9b9d2fc9d7306fb1e238b35d0c1]]
Former-commit-id: dd90c991389a2d6197da3ceea7b9da3969c28397
Former-commit-id: 65f8c2dbef7461538b2253bab0da86ab0318ca50