pack-objects: use free()+xcalloc() instead of xrealloc()+memset()
Whenever the hash table becomes too small then its size is increased,
the original part (and the added space) is zerod out using memset(),
and the table is rebuilt from scratch.

Simplify this proceess by returning the old memory using free() and
allocating the new buffer using xcalloc(), which already clears the
buffer for us.  That way we avoid copying the old hash table contents
needlessly inside xrealloc().

While at it, use the first array member with sizeof instead of a
specific type.  The old code used uint32_t and int, while index is
actually an array of int32_t.  Their sizes are the same basically
everywhere, so it's not actually a problem, but the new code is
cleaner and doesn't have to be touched should the type be changed.

Signed-off-by: Rene Scharfe <l.s.r@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8685e25d9a86169e4e145a3253e0fa57961fa348 [formerly 8685e25d9a86169e4e145a3253e0fa57961fa348 [formerly fb79947487b628ec82bc0ad46a08629539a59091]]
Former-commit-id: 90f00d6218e3cde1ab5f5790cc2568cf05bc05ae
Former-commit-id: 4b2d4f1e522c1c3326aa00522ea41e92968274b6