error_resolve_conflict: drop quotations around operation
When you try to commit with unmerged entries, you get an
error like:

  $ git commit
  error: 'commit' is not possible because you have unmerged files.

The quotes around "commit" are clunky; the user doesn't care
that this message is a template with the command-name filled
in.  Saying:

  error: commit is not possible because you have unmerged files

is easier to read. As this code is called from other places,
we may also end up with:

  $ git merge
  error: merge is not possible because you have unmerged files

  $ git cherry-pick foo
  error: cherry-pick is not possible because you have unmerged files

  $ git revert foo
  error: revert is not possible because you have unmerged files

All of which look better without the quotes. This also
happens to match the behavior of "git pull", which generates
a similar message (but does not share code, as it is a shell
script).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b22207ac275ff8960cc45cdad750df7015bdb26f [formerly 5550c2a88cd50859a70974cd65b93b5a7d22a5c3] [formerly 5550c2a88cd50859a70974cd65b93b5a7d22a5c3 [formerly d795216ac35f08e1ae2a706fca4c06a887c7d797]]
Former-commit-id: b60c581017a421a2868a9c2d2394db9d686646d7
Former-commit-id: 1cce534150325284125156193cb69531b842aeb5