string-list: spell all values out that are given to a string_list initializer
STRING_LIST_INIT_{NODUP,DUP} initializers list values only
for earlier structure members, relying on the usual
convention in C that the omitted members are initailized to
0, i.e. the former is expanded to the latter:

	struct string_list l = STRING_LIST_INIT_DUP;
	struct string_list l = { NULL, 0, 0, 1 };

and the last member that is not mentioned (i.e. 'cmp') is
initialized to NULL.

While there is nothing wrong in this construct, spelling out
all the values where the macros are defined will serve also
as a documentation, so let's do so.

Signed-off-by: Tanay Abhra <tanayabh@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b2ec4fbe5f6cbf754ae6387c84b14c5e6b556130 [formerly 5ae86ef10a6f369f19e0f914da7a14b3cd9edaea] [formerly 5ae86ef10a6f369f19e0f914da7a14b3cd9edaea [formerly acb3d22264fd3736f95f71aa38e1d6fa01f5e9e1]]
Former-commit-id: 368938d919b47ba25f8192cabdaf49a74870996f
Former-commit-id: 4edb749c5fac96498e6b92ec1f47387fa8854df0