refs.c: change read_ref_at to use the reflog iterators
read_ref_at has its own parsing of the reflog file for no really good reason
so lets change this to use the existing reflog iterators. This removes one
instance where we manually unmarshall the reflog file format.

Remove the now redundant ref_msg function.

Log messages for errors are changed slightly. We no longer print the file
name for the reflog, instead we refer to it as 'Log for ref <refname>'.
This might be a minor useability regression, but I don't really think so, since
experienced users would know where the log is anyway and inexperienced users
would not know what to do about/how to repair 'Log ... has gap ...' anyway.

Adapt the t1400 test to handle the change in log messages.

Signed-off-by: Ronnie Sahlberg <sahlberg@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c3343cbfecdb5b1841cc5b05a02afc415c2ca241 [formerly c3343cbfecdb5b1841cc5b05a02afc415c2ca241 [formerly 4207ed285f31ad3e04f08254237c0c1a1609642b]]
Former-commit-id: d38aaf7c7c519b6370e3213e4864a6efbff06664
Former-commit-id: e69114aa560d8db0447d6335af995c0fade3bd44