t5551: fix the 50,000 tag test
The first version of test 23 did simply check that no output was
sent to the standard error stream.  With 5e2c7cd2 (t5551: do not use
unportable sed '\+', 2013-05-12), we started to also verify that the
expected tags were actually cloned.

Since 68b939b2 (clone: send diagnostic messages to stderr,
2013-09-18), "git clone" shows "Cloning into 'too-many-refs'" to the
standard error stream (it used to do so to the standard output),
causing the test to fail.

Signed-off-by: Torsten Bögershausen <tboegi@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fff08544916c461b912e73f8b9e4539c7d5e730c [formerly dd27fae5f918c4256a82e4d6b35cd8b4fe7e0926] [formerly dd27fae5f918c4256a82e4d6b35cd8b4fe7e0926 [formerly 9e10e0b9a01cf2f611f1ef2330e27a0822cf3de6]]
Former-commit-id: 22c51ae6dfb22fec3b5f162ff9807bd4dde14331
Former-commit-id: 00ce22b7536a95a0b4e37eeff8b5c6903774400f