git potty: restore environments after alias expansion
Commit 4ad8332 (t0001: test git init when run via an alias -
2010-11-26) noted breakages when running init via alias. The problem
is for alias to be used, $GIT_DIR must be searched, but 'init' and
'clone' are not happy with that. So we start a new process like an
external command, with clean environment in this case. Env variables
that are set by command line (e.g. "git --git-dir=.. ") are kept.

This should also fix autocorrecting a command typo to "init" because
it's the same problem: aliases are read, then "init" is unhappy with
$GIT_DIR already set up because of that.

Reminded-by: David Turner <dturner@twopensource.com>
Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7e37f4da57354d94301a1a731e8191d5fb50fbae [formerly 484cc50715740e6c1706389cb04d48ae113382ff] [formerly 484cc50715740e6c1706389cb04d48ae113382ff [formerly c0562611c525bb3c564b79c345fc7e4f9e799e54]]
Former-commit-id: 671746d1604f413d5f48aa1b24898e4c7bbea0c5
Former-commit-id: 9b51ac496bc16438909b072517abb82743155587