t3302: drop unnecessary NOT_EXPENSIVE pseudo-prerequisite
This was only necessary because do_tests helper the script defines
took its parameters in a wrong order.  Just pass an empty string (or
not passing the optional EXPENSIVE prerequisite) when running the
test with a light-weight set of parameters and have the shell do the
right thing when parsing test_expect_success helper.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: feb1cbfbfca010ef711ee1bd4328f323c119320a [formerly 2715d5507fed10182a1ab34e9ee19a777a7120a9] [formerly 2715d5507fed10182a1ab34e9ee19a777a7120a9 [formerly 19c8c4a9b70a5c896d2c88250d7f209970fdf8d0]]
Former-commit-id: d6029ed61c52226ff7ba41eab5abf68dc0fe2046
Former-commit-id: 1c8040e249c715151250e44971f76f40fa80d3b9