t3419: drop unnecessary NOT_EXPENSIVE pseudo-prerequisite
This was only necessary because do_tests helper the script defines
took its parameters in a wrong order.  Just pass an empty string (or
not passing the optional EXPENSIVE prerequisite) when running the
test with a light-weight set of parameters and have the shell do the
right thing when parsing test_expect_success helper.

Also update coding style while we are at it.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 16ae8353071393dd63c4cb2e1471678a1474314d [formerly 4d104ab9060346f928525b7c9a77784e60550986] [formerly 4d104ab9060346f928525b7c9a77784e60550986 [formerly b687cd6aba394bba34ad9fb70aa33cd5cfcebc6c]]
Former-commit-id: 4b31b7bd3e7583d0b60d58fcacf609727c55bc77
Former-commit-id: 8b453534a410bb3890a556a6f678b49dcaa04d5a