repack: do not accidentally pack kept objects by default
Commit ee34a2b (repack: add `repack.packKeptObjects` config
var, 2014-03-03) added a flag which could duplicate kept
objects, but did not mean to turn it on by default. Instead,
the option is tied by default to the decision to write
bitmaps, like:

  if (pack_kept_objects < 0)
	  pack_kept_objects = write_bitmap;

after which we expect pack_kept_objects to be a boolean 0 or
1.  However, that assignment neglects that write_bitmap is
_also_ a tri-state with "-1" as the default, and with
neither option given, we accidentally turn the option on.

This patch is the minimal fix to restore the desired
behavior for the default state. Further patches will fix the
more complicated cases.

Note the update to t7700. It failed to turn on bitmaps,
meaning we were actually confirming the wrong behavior!

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b438f1870cab6562113625c47443b557ed109ca1 [formerly b438f1870cab6562113625c47443b557ed109ca1 [formerly 64d3dc94687bf4a8a422ab6d5cffcdb44315d49e]]
Former-commit-id: 80ebdb241d13b0733b74fa23d5faeb669b21ddf7
Former-commit-id: a7ada83b6dbe28e7ed8006f9d04ecc8080b44715