repack: respect pack.writebitmaps
The config option to turn on bitmaps is read all the way
down in the plumbing of pack-objects. This makes it hard for
other options in the porcelain of repack to make decisions
based on the bitmap setting. For example,
repack.packKeptObjects tries to kick in by default only when
bitmaps are turned on. But it can't do so reliably because
it doesn't yet know whether we are using bitmaps.

This patch teaches repack to respect pack.writebitmaps. It
means we pass a redundant command-line flag to pack-objects,
but that's OK; it shouldn't affect the outcome.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: cf5fc6a3b8e232cef72164000439b4c667400be1 [formerly cf5fc6a3b8e232cef72164000439b4c667400be1 [formerly 3198b89fb225ef8e0200e9486a48be8068e85d15]]
Former-commit-id: 2af4a27ab51e323d766d4e87e6264fb3497780e3
Former-commit-id: ec00fac5aaedfd4a511c8798cacdc88717dd434f