pack-objects: stop respecting pack.writebitmaps
The handling of the pack.writebitmaps config option
originally happened in pack-objects, which is quite
low-level. It would make more sense for drivers of
pack-objects to read the config, and then manipulate
pack-objects with command-line options.

Recently, repack learned to do so, making the low-level read
of pack.writebitmaps redundant here. Other callers, like
upload-pack, would not generally want to write bitmaps
anyway.

This could be considered a regression for somebody who is
driving pack-objects themselves outside of repack and
expects the config option to be used. However, such users
seem rather unlikely given how new the bitmap code is (and
the fact that they would basically be reimplementing repack
in the first place).

Note that we do not do anything with pack.writeBitmapHashCache
here. That option is not about "do we write bimaps", but
rather "when we are writing bitmaps, how do we do it?". You
would want that to kick in anytime you decide to write them,
similar to how pack.indexVersion is used.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6a308f66f04832c764ac5bed8aa2f49b44838e32 [formerly 5fbc122459e85b2eab0172ee3b46790374ebd56e] [formerly 5fbc122459e85b2eab0172ee3b46790374ebd56e [formerly 15a906c5e99fc34534fd40d61590da142dee71d2]]
Former-commit-id: 0c55f5d1abfc75f8b01f2b1aa1520745ff26457a
Former-commit-id: bdbcf254115d4f6858cc722dc349bbc4ea180da8