repack: simplify handling of --write-bitmap-index
We previously needed to pass --no-write-bitmap-index
explicitly to pack-objects to override its reading of
pack.writebitmaps from the config. Now that it no longer
does so, we can assume that bitmaps are off by default, and
only turn them on when necessary. This also lets us avoid a
confusing tri-state flag for write_bitmaps.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b7898aaa8bafa9aab9fc11b882ab76d6275f99b2 [formerly b7898aaa8bafa9aab9fc11b882ab76d6275f99b2 [formerly 2bed2d47b4394bd6d4ae4645be9f7424009d3c9c]]
Former-commit-id: ba8b3082493a2e469640836c96b5b90436150f06
Former-commit-id: 812307ed3aea47ea80993fb5c639c6860f85fd5b