replace dangerous uses of strbuf_attach
It is not a good idea to strbuf_attach an arbitrary pointer
just because a function you are calling wants a strbuf.
Attaching implies a transfer of memory ownership; if anyone
were to modify or release the resulting strbuf, we would
free() the pointer, leading to possible problems:

  1. Other users of the original pointer might access freed
     memory.

  2. The pointer might not be the start of a malloc'd
     area, so calling free() on it in the first place would
     be wrong.

In the two cases modified here, we are fortunate that nobody
touches the strbuf once it is attached, but it is an
accident waiting to happen.  Since the previous commit,
commit_tree and friends take a pointer/buf pair, so we can
just do away with the strbufs entirely.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: dee6f6fdfe643a7b37028fdb5451ff5fcd5ba080 [formerly dee6f6fdfe643a7b37028fdb5451ff5fcd5ba080 [formerly e6dfcd6767a58816dacec0df39515803b267fbe6]]
Former-commit-id: 3df145a54fd238ade08638469e59f65fe074b244
Former-commit-id: 1867202ba354d75f8be31155f68f8086243c5f13