commit: push commit_index update into alloc_commit_node
Whenever we create a commit object via lookup_commit, we
give it a unique index to be used with the commit-slab API.
The theory is that any "struct commit" we create would
follow this code path, so any such struct would get an
index. However, callers could use alloc_commit_node()
directly (and get multiple commits with index 0).

Let's push the indexing into alloc_commit_node so that it's
hard for callers to get it wrong.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 27a3b5a563e2dc1ba5ada2ea41ce89293575598b [formerly 27a3b5a563e2dc1ba5ada2ea41ce89293575598b [formerly 969eba6341a5af8ac52c67e26462548ed05e23e3]]
Former-commit-id: 59a352bbc0664ebfe9995b2454cac7fdf94a7972
Former-commit-id: 315adaf5e1bb900b477bb1b05b261959be3641df