do not create "struct commit" with xcalloc
In both blame and merge-recursive, we sometimes create a
"fake" commit struct for convenience (e.g., to represent the
HEAD state as if we would commit it). By allocating
ourselves rather than using alloc_commit_node, we do not
properly set the "index" field of the commit. This can
produce subtle bugs if we then use commit-slab on the
resulting commit, as we will share the "0" index with
another commit.

We can fix this by using alloc_commit_node() to allocate.
Note that we cannot free the result, as it is part of our
commit allocator. However, both cases were already leaking
the allocated commit anyway, so there's nothing to fix up.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6ff4e82dc1f555f0d72dab38fafaae59b4130226 [formerly 6ff4e82dc1f555f0d72dab38fafaae59b4130226 [formerly 10322a0aaf84382d8901f9ab59e59c39f0c035bb]]
Former-commit-id: 4d93060ac0310a55eacc2c5ce5c2729771750e25
Former-commit-id: e20e822349039b9f2c23709646bd993d5cde1dd6