logmsg_reencode: return const buffer
The return value from logmsg_reencode may be either a newly
allocated buffer or a pointer to the existing commit->buffer.
We would not want the caller to accidentally free() or
modify the latter, so let's mark it as const.  We can cast
away the constness in logmsg_free, but only once we have
determined that it is a free-able buffer.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 30339704a9fc6df352ff52ebea3b8dad5ea8b516 [formerly ffd275ef6a39853bf637a37d79dc6bc2af4c432b] [formerly ffd275ef6a39853bf637a37d79dc6bc2af4c432b [formerly b000c59b0c80fc187e5e0e48dc9396cd60576c4e]]
Former-commit-id: b9588bc32cced554acae6aac85967fbfe3908c9c
Former-commit-id: cc1d421b47a2f4f6114dfa684a9f4166798154b7