sequencer: use logmsg_reencode in get_message
This simplifies the code, as logmsg_reencode handles the
reencoding for us in a single call. It also means we learn
logmsg_reencode's trick of pulling the buffer from disk when
commit->buffer is NULL (we currently just silently return!).
It is doubtful this matters in practice, though, as
sequencer operations would not generally turn off
save_commit_buffer.

Note that we may be fixing a bug here. The existing code
does:

  if (same_encoding(to, from))
	  reencode_string(buf, to, from);

That probably should have been "!same_encoding".

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9d4b663e9543ac08ce5fcfd4db7dfbdeabcc7bae [formerly 25dd8698e5119bacd90b52012ec8628684f28487] [formerly 25dd8698e5119bacd90b52012ec8628684f28487 [formerly d74a4e57d2703e841deed994752b948a768f31ec]]
Former-commit-id: 8cc7cc4d6889c9778620a00993a95ff359800a4a
Former-commit-id: 64926970b132aa2df06afb7521858229e5824002