use get_cached_commit_buffer where appropriate
Some call sites check commit->buffer to see whether we have
a cached buffer, and if so, do some work with it. In the
long run we may want to switch these code paths to make
their decision on a different boolean flag (because checking
the cache may get a little more expensive in the future).
But for now, we can easily support them by converting the
calls to use get_cached_commit_buffer.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ff38a257fcada9b75a01bffd5efd1c2d452fb61d [formerly ff38a257fcada9b75a01bffd5efd1c2d452fb61d [formerly a97934d8205772ffd2a528a9e970af7dec725012]]
Former-commit-id: d426d8803cf4f138243b4664bb8ea7822eeb84e1
Former-commit-id: 12f1cc1ae9cde94e1ed2ba5ecb8d301ab194b47d