convert logmsg_reencode to get_commit_buffer
Like the callsites in the previous commit, logmsg_reencode
already falls back to read_sha1_file when necessary.
However, I split its conversion out into its own commit
because it's a bit more complex.

We return either:

  1. The original commit->buffer

  2. A newly allocated buffer from read_sha1_file

  3. A reencoded buffer (based on either 1 or 2 above).

while trying to do as few extra reads/allocations as
possible. Callers currently free the result with
logmsg_free, but we can simplify this by pointing them
straight to unuse_commit_buffer. This is a slight layering
violation, in that we may be passing a buffer from (3).
However, since the end result is to free() anything except
(1), which is unlikely to change, and because this makes the
interface much simpler, it's a reasonable bending of the
rules.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3ebfa87c710ae369ff380788f79184629bccee5e [formerly 3ebfa87c710ae369ff380788f79184629bccee5e [formerly b66103c3baa593a39b8b0751213b9fce60e94de4]]
Former-commit-id: 0cbb5e78ee776aa53c05146b3e1bc417496bbbe0
Former-commit-id: 041b650d46ee2f28432b6537c8e0c3e63b69f736