use get_commit_buffer everywhere
Each of these sites assumes that commit->buffer is valid.
Since they would segfault if this was not the case, they are
likely to be correct in practice. However, we can
future-proof them by using get_commit_buffer.

And as a side effect, we abstract away the final bare uses
of commit->buffer.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 09a71b39a5bccf3d7bd0a4b0d0952ed8150d7c66 [formerly 09a71b39a5bccf3d7bd0a4b0d0952ed8150d7c66 [formerly bc6b8fc1300ef79c4b4c3c2a79bb3c1e2e032963]]
Former-commit-id: 830cf793f90c693d3c1898cbe95fa0e41a909a3c
Former-commit-id: 6ad74527b32be19fa3c9eb649a37312e52e8fc4b