commit: convert commit->buffer to a slab
This will make it easier to manage the buffer cache
independently of the "struct commit" objects. It also
shrinks "struct commit" by one pointer, which may be
helpful.

Unfortunately it does not reduce the max memory size of
something like "rev-list", because rev-list uses
get_cached_commit_buffer() to decide not to show each
commit's output (and due to the design of slab_at, accessing
the slab requires us to extend it, allocating exactly the
same number of buffer pointers we dropped from the commit
structs).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 360fe967a7a9b663e8d6f40f27e27330b881866b [formerly 360fe967a7a9b663e8d6f40f27e27330b881866b [formerly c1b3c71f4b4571abb2b2a457122fd100dc9f7eb0]]
Former-commit-id: 9b96d82fc890772ec9e37d68a225541f6c8e077b
Former-commit-id: 37f039e576ccc15b6b53c770c75571bca831ad85