commit: record buffer length in cache
Most callsites which use the commit buffer try to use the
cached version attached to the commit, rather than
re-reading from disk. Unfortunately, that interface provides
only a pointer to the NUL-terminated buffer, with no
indication of the original length.

For the most part, this doesn't matter. People do not put
NULs in their commit messages, and the log code is happy to
treat it all as a NUL-terminated string. However, some code
paths do care. For example, when checking signatures, we
want to be very careful that we verify all the bytes to
avoid malicious trickery.

This patch just adds an optional "size" out-pointer to
get_commit_buffer and friends. The existing callers all pass
NULL (there did not seem to be any obvious sites where we
could avoid an immediate strlen() call, though perhaps with
some further refactoring we could).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 220d673995ed2f3ecd73246bde2fc7368b10bad9 [formerly b348ff7b1206542113e624a11d669effb8f9486d] [formerly b348ff7b1206542113e624a11d669effb8f9486d [formerly 8597ea3afea067b39ba7d4adae7ec6c1ee0e7c91]]
Former-commit-id: d13ff3f90eb0c1fdfda0d4106553c5845921fe52
Former-commit-id: 485d786fd01b1e269c77889689aa479d9b343ecb