read-cache: store in-memory flags in the first 12 bits of ce_flags
We're running out of room for in-memory flags. But since b60e188
(Strip namelen out of ce_flags into a ce_namelen field - 2012-07-11),
we copy the namelen (first 12 bits) to ce_namelen field. So those bits
are free to use. Just make sure we do not accidentally write any
in-memory flags back.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7689fcba85160e3804bceb566c975d1cefbb8760 [formerly 7689fcba85160e3804bceb566c975d1cefbb8760 [formerly ce51bf09f8f0de70673b7cbdbc0e0df0b07e1def]]
Former-commit-id: 74fb51a81b3087b0c09f003d73617f4f7e88ea76
Former-commit-id: 333c97f445bb8a6568b99f18e93cf4607d665a62