split-index: strip pathname of on-disk replaced entries
We know the positions of replaced entries via the replace bitmap in
"link" extension, so the "name" path does not have to be stored (it's
still in the shared index). With this, we also have a way to
distinguish additions vs replacements at load time and can catch
broken "link" extensions.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a75d82acca2dbb95d7feb404decb225b92203a6a [formerly a75d82acca2dbb95d7feb404decb225b92203a6a [formerly b3c96fb158f05152336f167076f5d81d23c3a5e5]]
Former-commit-id: 97c56a64b5312abbbacf412f3d4b22b9c3d39e7c
Former-commit-id: 2ff1f81c71c13c45f48f54cc1b576f8a10752b26