Win32: reliably detect console pipe handles
As of "Win32: Thread-safe windows console output", child processes may
print to the console even if stdout has been redirected to a file. E.g.:

 git config tar.cat.command "cat"
 git archive -o test.cat HEAD

Detecting whether stdout / stderr point to our console pipe is currently
based on the assumption that OS HANDLE values are never reused. This is
apparently not true if stdout / stderr is replaced via dup2() (as in
builtin/archive.c:17).

Instead of comparing handle values, check if the file descriptor isatty()
backed by a pipe OS handle. This is only possible by swapping the handles
in MSVCRT's internal data structures, as we do in winansi_init().

Reported-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Karsten Blees <blees@dcon.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 859b1a9ae8abfdf0b317228dadaa33d0540fb77f [formerly 859b1a9ae8abfdf0b317228dadaa33d0540fb77f [formerly 51822653f58e6b9b6119f560ff864813fbde81ba]]
Former-commit-id: 4548e8ccbf6f9cedbad0c19faefa2f2c917e7885
Former-commit-id: 65c24a8a1aeb13515301aa7572909936c2aaf6c2