merge: add t7613 for submodule updates
Test that the merge command updates the work tree as expected (for
submodule changes which don't result in conflicts) when used without
arguments or with the '--ff', '--ff-only' and '--no-ff' flag.

Implement the KNOWN_FAILURE_NOFF_MERGE_DOESNT_CREATE_EMPTY_SUBMODULE_DIR
switch to expect the known failure that --no-ff merges do not create the
empty submodule directory.

The KNOWN_FAILURE_NOFF_MERGE_ATTEMPTS_TO_MERGE_REMOVED_SUBMODULE_FILES
switch is also implemented to expect the known failure that --no-ff
merges attempt to merge the new files in the former submodule directory
with those of the removed submodule.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 528b86e949f209e5a45cff040486864b50e1813f [formerly 06548bf546f9d4de91f7d7234f3033b983f58693] [formerly 06548bf546f9d4de91f7d7234f3033b983f58693 [formerly 663ed39a889dae8beee2fc6953607f7660a5cddf]]
Former-commit-id: 7d4791f189b099ccf544696c228f85ad91e53414
Former-commit-id: 9de93c1b64e9e74907b9c769c47ef628825e3839