pull: add t5572 for submodule updates
Test that the pull command updates the work tree as expected (for
submodule changes which don't result in conflicts) when used without
arguments or with the '--ff', '--ff-only' and '--no-ff' flag each. Add
helper functions to reset the branch to be updated to to the current
HEAD so that pull is doing the transition from HEAD to the given branch.

Set KNOWN_FAILURE_NOFF_MERGE_ATTEMPTS_TO_MERGE_REMOVED_SUBMODULE_FILES
and KNOWN_FAILURE_NOFF_MERGE_DOESNT_CREATE_EMPTY_SUBMODULE_DIR to
document that pull has the same --no-ff known failures merge has.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 75b9ee8e5c350f1691ce83908647d6510a3942f2 [formerly 75b9ee8e5c350f1691ce83908647d6510a3942f2 [formerly 921f50b48e2d5283bbd3c5ea2c64ad7407a2bc1c]]
Former-commit-id: 4ea0d6f25219e0d28c450d16085bf6684baf8cd0
Former-commit-id: 807d150c16f44b3812508150cd3dd597387fa763