am: add t4255 for submodule updates
Test that the am command updates the work tree as expected (for submodule
changes which don't result in conflicts). To make that work add two
helper functions that use format-patch to create the input for am.

Add the KNOWN_FAILURE_NOFF_MERGE_ATTEMPTS_TO_MERGE_REMOVED_SUBMODULE_FILES
switch to expect the known failure that --no-ff merges attempt to merge
the new files in the former submodule directory with those of the removed
submodule.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 059abbd421d715b3477816c2471f2b10af000611 [formerly 059abbd421d715b3477816c2471f2b10af000611 [formerly 23e2f388c7da3cfe72c880e662dafb0c2965abba]]
Former-commit-id: 3bcf4827a660cb847cd958aaa2fe2ed4d0b875d4
Former-commit-id: ba86829be900e9d09f5914bc83ab24208abe25d9