rebase--merge: fix --skip with two conflicts in a row
If git rebase --merge encountered a conflict, --skip would not work if the
next commit also conflicted.  The msgnum file would never be updated with
the new patch number, so no patch would actually be skipped, resulting in an
inescapable loop.

Update the msgnum file's value as the first thing in call_merge.  This also
avoids an "Already applied" message when skipping a commit.  There is no
visible change for the other contexts in which call_merge is invoked, as the
msgnum file's value remains unchanged in those situations.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: afac771f83fbfef44506af54b01d17a7f39a2ebe [formerly afac771f83fbfef44506af54b01d17a7f39a2ebe [formerly 95104c7e257652b82aed089494def344e3938928]]
Former-commit-id: e61db60a4d04b768600309cc9bf89a791fc4522c
Former-commit-id: d4b7fa1b7f94d23752e6e1b142245e56ab3a1553