t7510: stop referring to master in later tests
Our setup creates a sequence of commits, each with its own
tag. However, we sometimes refer to "seventh-signed" as
"master". This works, since it is at the tip of the created
branch, but is brittle if new tests need to add more
commits. Let's use its tag name to be unambiguous.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a48707ff9ddbd8a3ec7de2d083c3bf7ab203776c [formerly a48707ff9ddbd8a3ec7de2d083c3bf7ab203776c [formerly 526d56e07200483fc93be9c2dfee81bb87713c9b]]
Former-commit-id: 041fc1ac582f9b03a9fe25f2eff980357a3625ec
Former-commit-id: b5659a4c973cc44a24c64730408de28401168b53