pretty: avoid reading past end-of-string with "%G"
If the user asks for --format=%G with nothing else, we
correctly realize that "%G" is not a valid placeholder (it
should be "%G?", "%GK", etc). But we still tell the
strbuf_expand code that we consumed 2 characters, causing it
to jump over the trailing NUL and output garbage.

This also fixes the case where "%GX" would be consumed (and
produce no output). In other cases, we pass unrecognized
placeholders through to the final string.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9a2ee5dd5536667836ef2225d7d19f4519512ff8 [formerly 9a2ee5dd5536667836ef2225d7d19f4519512ff8 [formerly aa4b78d483a918ebee810993e420b4697b0de4d3]]
Former-commit-id: 8201e1c3c982b748daf77efab55d0bd9f1ba3f5e
Former-commit-id: 3e862d388e89e6a3f4245536903ba5c20ff23b21