apply: use skip_prefix instead of raw addition
A submodule diff generally has content like:

  -Subproject commit [0-9a-f]{40}
  +Subproject commit [0-9a-f]{40}

When we are using "git apply --index" with a submodule, we
first apply the textual diff, and then parse that result to
figure out the new sha1.

If the diff has bogus input like:

  -Subproject commit 1234567890123456789012345678901234567890
  +bogus

we will parse the "bogus" portion. Our parser assumes that
the buffer starts with "Subproject commit", and blindly
skips past it using strlen(). This can cause us to read
random memory after the buffer.

This problem was unlikely to have come up in practice (since
it requires a malformed diff), and even when it did, we
likely noticed the problem anyway as the next operation was
to call get_sha1_hex on the random memory.

However, we can easily fix it by using skip_prefix to notice
the parsing error.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 55a81e4c0cbb4f6e1ea4fdc69859a2b903194f6b [formerly 4a3ad7a92837572b780ed464760f31cdb55f2f56] [formerly 4a3ad7a92837572b780ed464760f31cdb55f2f56 [formerly ce2ecf29241ca90655da37918044cbbaff48a2bd]]
Former-commit-id: aa32c38a63e92f4704b1a70bf9594411faaa7b48
Former-commit-id: 882b3b452cb9f390593f8335cc6a674448222f04