fast-import: fix read of uninitialized argv memory
Fast-import shares code between its command-line parser and
the "option" command. To do so, it strips the "--" from any
command-line options and passes them to the option parser.
However, it does not confirm that the option even begins
with "--" before blindly passing "arg + 2".

It does confirm that the option starts with "-", so the only
affected case was:

  git fast-import -

which would read uninitialized memory after the argument. We
can fix it by using skip_prefix and checking the result. As
a bonus, this gets rid of some magic numbers.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2916c2f91416e1c72be1041095ece2d56ef1ecbd [formerly 65de851879b326844e3930bfd9635bdc677ead61] [formerly 65de851879b326844e3930bfd9635bdc677ead61 [formerly ff45c0d4a316d620d118ec628dd8e78597a23321]]
Former-commit-id: 4617dc95446896539d287877082c9a3be34b1081
Former-commit-id: e09836c1fdc7a8a371b6d520927dfbe33cd9666b