transport-helper: avoid reading past end-of-string
We detect the "import-marks" capability by looking for that
string, but _without_ a trailing space. Then we skip past it
using strlen("import-marks "), with a space. So if a remote
helper gives us exactly "import-marks", we will read past
the end-of-string by one character.

This is unlikely to be a problem in practice, because such
input is malformed in the first place, and because there is
a good chance that the string has an extra NUL terminator
one character after the original (because it formerly had a
newline in it that we parsed off).

We can fix it by using skip_prefix with "import-marks ",
with the space. The other form appears to be a typo from
a515ebe (transport-helper: implement marks location as
capability, 2011-07-16); "import-marks" has never existed
without an argument, and it should match the "export-marks"
definition above.

Speaking of which, we can also use skip_prefix in a few
other places while we are in the function.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 75ad5c22c57b22248a2b3193a1defdeae158bca6 [formerly 1f2e0ffa8afcf123a1716796a90b54fb87fdcac1] [formerly 1f2e0ffa8afcf123a1716796a90b54fb87fdcac1 [formerly 21a2d4ada52132e6b0b67f8e28aa4bcda416f7f2]]
Former-commit-id: 8335942fb57d318759f7a854ecbdb21e809a8665
Former-commit-id: 626d443000e59ce8ec4478b4c8c26048d380fe11