use skip_prefix to avoid repeating strings
It's a common idiom to match a prefix and then skip past it
with strlen, like:

  if (starts_with(foo, "bar"))
	  foo += strlen("bar");

This avoids magic numbers, but means we have to repeat the
string (and there is no compiler check that we didn't make a
typo in one of the strings).

We can use skip_prefix to handle this case without repeating
ourselves.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4f3a282b935dfba2be7c5a9f94e5d5c6379f4db6 [formerly 4f3a282b935dfba2be7c5a9f94e5d5c6379f4db6 [formerly 95b567c7c3cf6b85d74b79424cdfbd40a7dee7c9]]
Former-commit-id: a22084316f2d8e547a0ea2b3cd640cbbac142feb
Former-commit-id: 38956e5922612e38c6f33bdd3fe9240c4be19196