fast-import: use skip_prefix for parsing input
Fast-import does a lot of parsing of commands and
dispatching to sub-functions. For example, given "option
foo", we might recognize "option " using starts_with, and
then hand it off to parse_option() to do the rest.

However, we do not let parse_option know that we have parsed
the first part already. It gets the full buffer, and has to
skip past the uninteresting bits. Some functions simply add
a magic constant:

  char *option = command_buf.buf + 7;

Others use strlen:

  char *option = command_buf.buf + strlen("option ");

And others use strchr:

  char *option = strchr(command_buf.buf, ' ') + 1;

All of these are brittle and easy to get wrong (especially
given that the starts_with call and the code that assumes
the presence of the prefix are far apart). Instead, we can
use skip_prefix, and just pass each handler a pointer to its
arguments.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: af5b03e53af2afee8e7be104da8452086a50d6bd [formerly af5b03e53af2afee8e7be104da8452086a50d6bd [formerly 97313bef2a16eb8b3d40830204c20c34ba9d6554]]
Former-commit-id: c867a6a6d946072c99e999fe3d0c474528ead18f
Former-commit-id: 7a350e2a16b15d3a91b2750eb8b164eb4cd6b559