daemon: use skip_prefix to avoid magic numbers
Like earlier cases, we can use skip_prefix to avoid magic
numbers that must match the length of starts_with prefixes.
However, the numbers are a little more complicated here, as
we keep parsing past the prefix. We can solve it by keeping
a running pointer as we parse; its final value is the
location we want.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6a73944188f771a4620c5968e01a1c83a60e2334 [formerly 6a73944188f771a4620c5968e01a1c83a60e2334 [formerly d12c24d2a96f8f9aeb100b800d8fb217f28a6a2a]]
Former-commit-id: dc234122ea7a7e40ee5ede938cc390dc40782a85
Former-commit-id: 60fe2d2b04a32ea93fb2d2a185485d939aae1767