stat_opt: check extra strlen call
As in earlier commits, the diff option parser uses
starts_with to find that an argument starts with "--stat-",
and then adds strlen("stat-") to find the rest of the
option.

However, in this case the starts_with and the strlen are
separated across functions, making it easy to call the
latter without the former. Let's use skip_prefix instead of
raw pointer arithmetic to catch such a case.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e615827840b1b8a2b20f75493ba9d4d70ef42d62 [formerly 20937618d9f6f9b1915e861b62e5cba865db397d] [formerly 20937618d9f6f9b1915e861b62e5cba865db397d [formerly 0539cc0038ff3411da1fea342b7d6615643bff5b]]
Former-commit-id: 624770e68a7a2005eba7b0a075f78ac4e694ce49
Former-commit-id: ea4c31691f553840cd04d5724f053d00de3231e9