fast-import: refactor parsing of spaces
When we see a file change in a commit, we expect one of:

  1. A mark.

  2. An "inline" keyword.

  3. An object sha1.

The handling of spaces is inconsistent between the three
options. Option 1 calls a sub-function which checks for the
space, but doesn't parse past it. Option 2 parses the space,
then deliberately avoids moving the pointer past it. Option
3 detects the space locally but doesn't move past it.

This is confusing, because it looks like option 1 forgets to
check for the space (it's just buried). And option 2 checks
for "inline ", but only moves strlen("inline") characters
forward, which looks like a bug but isn't.

We can make this more clear by just having each branch move
past the space as it is checked (and we can replace the
doubled use of "inline" with a call to skip_prefix).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 04b319bd900d86ff674a2e3b156d1953198c4863 [formerly b8dd5d9ed40de734a452c3be23465008b8733dfa] [formerly b8dd5d9ed40de734a452c3be23465008b8733dfa [formerly e814c39c2fe7cc915ba70c0aa6f03156a28920fc]]
Former-commit-id: 05f9afe79493fed1c98ada79c759db28b54d561a
Former-commit-id: cf7d44d8ab1d5a11641ae319f29e5c3f79f02df0