fetch-pack: refactor parsing in get_ack
There are several uses of the magic number "line+45" when
parsing ACK lines from the server, and it's rather unclear
why 45 is the correct number. We can make this more clear by
keeping a running pointer as we parse, using skip_prefix to
jump past the first "ACK ", then adding 40 to jump past
get_sha1_hex (which is still magical, but hopefully 40 is
less magical to readers of git code).

Note that this actually puts us at line+44. The original
required some character between the sha1 and further ACK
flags (it is supposed to be a space, but we never enforced
that). We start our search for flags at line+44, which
meanas we are slightly more liberal than the old code.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 554a41847b7474eb5939561819d5ab1341649eda [formerly cf6928ee491493f139adfa2ad5b67dcd64fbff84] [formerly cf6928ee491493f139adfa2ad5b67dcd64fbff84 [formerly 82e56767aa9334bb0c4cfe81964a576778b93d6e]]
Former-commit-id: 76e341e446d4f4486dfba1f3782c1b82542c2804
Former-commit-id: 401822a3447477fcd726bcf25698d51336c9d512