git: avoid magic number with skip_prefix
After handling options, any leftover arguments should be
commands. However, we pass through "--help" and "--version",
so that we convert them into "git help" and "git version"
respectively.

This is a straightforward use of skip_prefix to avoid a
magic number, but while we are there, it is worth adding a
comment to explain this otherwise confusing behavior.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 53b851cd89930c40c410e02883f4b0ba914c087a [formerly e60d40a2a2fafd5749235295923f110c32cae42a] [formerly e60d40a2a2fafd5749235295923f110c32cae42a [formerly 6d87780399e1196932cd5ea72f84aa90224e7402]]
Former-commit-id: 601a808d1dbf3571e76834ed264ae74568d21fdc
Former-commit-id: a7b969785e6e81a1e6d90ecfdd4fdd76d1c7c8ec