use skip_prefix to avoid repeated calculations
In some cases, we use starts_with to check for a prefix, and
then use an already-calculated prefix length to advance a
pointer past the prefix. There are no magic numbers or
duplicated strings here, but we can still make the code
simpler and more obvious by using skip_prefix.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c81965b532cad4294e45a37bfff76af19c406f70 [formerly 13cd68fa50cdc011ca25b7d59028d6d96bb5b26a] [formerly 13cd68fa50cdc011ca25b7d59028d6d96bb5b26a [formerly de8118e153c5e527263086605e437ccca5d4f1ef]]
Former-commit-id: 9e3e8ad967ecd7bcb22cc8201a393b9e05dbf5ef
Former-commit-id: 0ddf8a7698bdcef5c78c0e3b1aad5965a1ebcf7f