strbuf: add xstrfmt helper
You can use a strbuf to build up a string from parts, and
then detach it. In the general case, you might use multiple
strbuf_add* functions to do the building. However, in many
cases, a single strbuf_addf is sufficient, and we end up
with:

  struct strbuf buf = STRBUF_INIT;
  ...
  strbuf_addf(&buf, fmt, some, args);
  str = strbuf_detach(&buf, NULL);

We can make this much more readable (and avoid introducing
an extra variable, which can clutter the code) by
introducing a convenience function:

  str = xstrfmt(fmt, some, args);

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7112155fbe403ae890ff2dd2893d524e8e8608de [formerly 7112155fbe403ae890ff2dd2893d524e8e8608de [formerly 30a0ddb705678d512185e359831479a6b3567147]]
Former-commit-id: 9c25fbb247def16f3710564c86bca3923c16a9a3
Former-commit-id: 27163ee15926794391a0b3f86e642584751424df