bisect: add t6041 for submodule updates
Test that the bisect command updates the work tree as expected. To make
that work with the new submodule test framework a git_bisect helper
function is added. This adds a commit after the one given to be switched
to and makes that one the bad commit. The starting point is then given to
bisect as the good commit which makes bisect change the work tree to the
commit in between, which is the commit given.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f533953f7194ccb6d28727464f721b23f2da5254 [formerly eb890e30e98247afe99408026d3d32e6ece50bb6] [formerly eb890e30e98247afe99408026d3d32e6ece50bb6 [formerly 8f8ba56b5b6321b07eba29bb7ae6a94b5ae0d34a]]
Former-commit-id: 5136d74b4403c24afca68fa6eb171b18a32cc6be
Former-commit-id: c7ffedd7dca6c209d40b0df4e36f9a4e4b11f053