rebase: add t3426 for submodule updates
Test that the rebase command updates the work tree as expected for
changes which don't result in conflicts. To make that work add two
helper functions that add a commit only touching files and then
revert it. This allows to rebase the target commit over these two
and to compare the result.

Set KNOWN_FAILURE_NOFF_MERGE_DOESNT_CREATE_EMPTY_SUBMODULE_DIR to
document that "replace directory with submodule" fails for an
interactive rebase because a directory "sub1" already exists.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 630fc38869984476bd783f03d4a9a013e1f82b22 [formerly 630fc38869984476bd783f03d4a9a013e1f82b22 [formerly c7e69168cf9215f50c42666c7e0dbb1f21c69e20]]
Former-commit-id: 04d09f98398f83a2dfcfa481bdbfc8554b119ea3
Former-commit-id: 4e71d6086d45a20a8e5d6a1c05fcfe2542e33292