stash: add t3906 for submodule updates
Test that the stash apply command updates the work tree as expected for
changes which don't result in conflicts. To make that work add a helper
function that uses read-tree to apply the changes of the target commit
to the work tree, then stashes these changes and at last applies that
stash.

Implement the KNOWN_FAILURE_STASH_DOES_IGNORE_SUBMODULE_CHANGES switch
and reuse two other already present switches to expect the known
failure that stash does ignore submodule changes.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: bd664079260fc3407d8694993f16ee502a6c7257 [formerly bd664079260fc3407d8694993f16ee502a6c7257 [formerly da7fe3fb6d810507bef586e76a01bb34758d4472]]
Former-commit-id: 722527c9e51edec14ca30f56e59155747773069e
Former-commit-id: 9732d05abba686df4c41504a30c8d2c955a331a7