use xstrfmt to replace xmalloc + strcpy/strcat
It's easy to get manual allocation calculations wrong, and
the use of strcpy/strcat raise red flags for people looking
for buffer overflows (though in this case each site was
fine).

It's also shorter to use xstrfmt, and the printf-format
tends to be easier for a reader to see what the final string
will look like.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c640214ff7ae4b1c227b9ae2991914a89dc77af1 [formerly c640214ff7ae4b1c227b9ae2991914a89dc77af1 [formerly b2724c87872aaec55dd7e5529aa029c3108b43a5]]
Former-commit-id: abbf75daa4b349b9a3f06b3e2642ea22579e8773
Former-commit-id: 7b0630ebd20514418a8acf174ef561f87e52a6a1